% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{APC}
\alias{APC}
\title{Absolute Proportional Change}
\usage{
APC(x, b = 0, l = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{b}{choose -1 (or less) to look into the future}

\item{l}{observations backwards}

\item{...}{dots passed}
}
\description{
\preformatted{
This is most useful for calculating velocity
and acceleration, (and jerk).
To get acceleration and jerk use this with
diffXts with differences 2 (and 3) respectively.

Ninety-nine (99%) percent of the people in the world
SHOULD HAVE BEEN using this one.
}
}
\examples{
\dontrun{

# APS(Absolute Proportional Change) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

APC(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1apc.0.1 V2apc.0.1
1970-01-01        NA        NA
1970-01-02        -3         1
1970-01-03        -1         1
}
}
