% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{buildModel}
\alias{buildModel}
\alias{buildModel.train}
\title{Build quantmod model given specified fitting method}
\usage{
buildModel(x, method, training.per, ...)

buildModel.train(quantmod, training.data, ...)
}
\arguments{
\item{x}{An object of class quantmod.  Required.  Created with specifyModel or an R formula}

\item{training.per}{Default is none. Required. Default will train on all data provided in the paramter object "quantmod".  Character vector representing dates in ISO 8601 format “CCYY-MM-DD” or “CCYY-MM-DD HH:MM:SS” of length 2.  training.per can also be a list of training dates (THE LIST FEATURE NOT YET IMPLMENTED).}

\item{...}{Dots.  Additional arguments to method call.}

\item{method_train}{Default is none. Defaults to "xgboost". A character string naming the R CRAN package caret function train fitting method.}
}
\value{
An object of class quantmod with fitted model attached.
}
\description{
Construct and attach a fitted model of type method to quantmod object.
This is a re-implementation of the R CRAN package buildModel with added flexibility in training.per

\preformatted{

# xgboost parameters
# https://xgboost.readthedocs.io/en/latest/parameter.html

# tuneGrid (production)

tg <- expand.grid(
 nrounds   =  c(500,200,100),
 eta       =  c(0.3,0.1,0.01,0.001),
 max_depth =  c(4,6,8,10),
 gamma     =  0,
 colsample_bytree = c(1,0.5),
 min_child_weight = 1,
 subsample        = c(1,0.5)
)

# tuneGrid (non-production)

tg <- expand.grid(
  nrounds   =  50,
  eta       =  c(0.1,0.01),
  max_depth =  c(4,7,10),
  gamma     =  0,
  colsample_bytree = 1,
  min_child_weight = 1,
  subsample        = 1
)

tc <- caret::trainControl(method = "cv", number = 5, allowParallel = T)

# of variable "specmodel", see
? econModel::specifyModel

# built model
specmodel <- buildModel(specmodel, method="train"
  , training.per=c("1970-12-31","2006-12-31")
  , method_train = "xgbTree", tuneGrid = tg, trControl = tc)
}
}
\examples{
\dontrun{
# of variable "specmodel", see
? econModel::specifyModel

# built model
specmodel <- buildModel(specmodel, method="train"
  , training.per=list(c("1970-12-31","1998-12-31"), c("2001-0101","2008-06-30"))
}
}
