% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSorting.R
\name{cAppend}
\alias{cAppend}
\alias{cAppend.default}
\alias{cAppend.data.frame}
\title{Custom Inserting/Appending Into R Objects}
\usage{
cAppend(x, values, ...)

\method{cAppend}{default}(x, values, ...)

\method{cAppend}{data.frame}(x, values, ...)
}
\arguments{
\item{x}{The R object, such that the values are to be inserted/appended to.}

\item{values}{List. The values to be included in the modified R object.}

\item{...}{Dots passed.}
}
\value{
Modified R Object.
}
\description{
This is a generalization of the R base package function "append".
}
\details{
Insert/append elements. If the elements are part of a collection, then the length of all elements of the list must be the same.
Currently, (at this time), supported, are R objects: "list", "vector" and class "data.frame".
"values" should be passed as a list. The new collection item names are taken from the List item names.

The type conversion function "ValuesFunction" can (if supported by the S3 method), transform the data:
ValueFunction = "Rfunction(value)".
}
\examples{
\dontrun{
cAppend(letters, list(c("zz","yy")),
  ValueFunction = "toupper(value)"
)
cAppend(iris[1:2,],
  list(END = c("zz","yy"), BEGIN = c("aa","bb")),
  ValueFunction = "toupper(value)", after = 3
  )
cAppend(as.list(iris[1:2,]),
  list(END = c("zz","yy"), BEGIN = c("aa","bb")),
  ValueFunction = "toupper(value)", after = 3
  )
}
}
