% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{caroline__dbWriteTable2}
\alias{caroline__dbWriteTable2}
\title{Using a Local data.frame Update a Currenly Existing PostgreSQL server table}
\usage{
caroline__dbWriteTable2(
  con,
  table.name,
  df,
  PartitionedCols,
  PartitionedColsValues,
  PrimaryKeyCols,
  IndexedCols,
  fill.null = TRUE,
  row.names = FALSE,
  display = TRUE,
  exec = TRUE,
  ...
)
}
\arguments{
\item{con}{connection.}

\item{table.name}{The name of the table to which the data frame is to be loaded.}

\item{df}{A dataframe to be loaded to the database.}

\item{PartitionedCols}{Vector of Strings.  If present, then of the [to be] partitioned table, this is the vector column names (in order). NOT YET IMPLEMENTED.}

\item{PartitionedColsValues}{List of Length One(1) R objects.  If present, then of the [to be] partitioned table partition values, this is the the list of column values (in order). NOT YET IMPLEMENTED.}

\item{PrimaryKeyCols}{Vector of Strings.  If present, then of the [to be] primary-keyed table, this is the the vector values (in order) are the primary-keyed columns. NOT YET IMPLEMENTED.}

\item{IndexedCols}{List of "Vector of Strings", or just one Vector of Strings.  If present, then of the [to be] indexed table, this is the the vector values (in order) are the indexed columns. NOT YET IMPLEMENTED.}

\item{fill.null}{Should new db present fields be added to the data.frame before it is loaded?}

\item{row.names}{Should the row names be loaded as a seperate column? (unlike the original dbWriteTable, default is FALSE)}

\item{display}{Logical. Whether to display the query ALTER TABLE ADD (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query ALTER TABLE ADD (defaults to \code{TRUE}).}
}
\value{
results of DBI::dbWriteTable
}
\description{
\preformatted{
This is R cran package "caroline" function "dbWriteTable2" with modifications
JAN 2021
https://github.com/cran/caroline/blob/af201137e4a31d675849291a1c9c07a0933b85de/R/database.R

modifications

1. add.id domain specific feature is removed
2. pg.update.seq domain specific feature is removed
3. "ORDER BY id DESC LIMIT 1" is replaced by " WHERE FALSE "
4. function "names" upon "df" is replaced with the modern "colnames"
   where appropriate
5. function "names" is replaced with the safer "Names"
6. "nv" is replaced with "nameVect".
7. Calls to DBI are called through "PACKAGE::".
   These are the (working) dbListFields, dbSendQuery, dbColumnInfo,
   dbClearResult, and dbWriteTable
8. If a column exists in the local table.name(data.frame) but does not
   exist on the server
   then it is added to the server's table fields.
9. Fix and Put back the type mismatch test.
10. Added drop = F to prevent dropping to a vector.
11. Changed message - not in fields of '", table.name,"' table. Omiting
                 To - not in fields of '", table.name,"' Adding.

For consistency and understandability renamed db.col.info to fields.info.
For consistency and understandability renamed db.precisions to fields.precisions.
For consistency and understandability renamed db.sclasses to fields.sclasses.
Since static, moved closer to the top calls to DBI::dbSendQuery, DBI::dbColumnInfo and DBI::dbClearResult
Moved other non-changing code to near the top.
}
}
