% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{clRemoveOldWorkDbasesEM}
\alias{clRemoveOldWorkDbasesEM}
\title{From the Cluster, Remove Old Work Databases}
\usage{
clRemoveOldWorkDbasesEM(display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{Dots}{passed.}
}
\value{
TRUE(always).  Will attempt to drop each 'rtmp%' database (but not the current work database).
}
\description{
In the cluster, logs on to the database "r_user_econmodel" as user "rtmp%" and (tries to) drop the "rtmp%" database.
Does not drop the "current work database": "getOption("econmodel_db_dbname")".
}
\details{
Note, is also attempt to drop the Old User.
}
\examples{
\dontrun{
clRemoveOldWorkDbasesEM()
}
}
