% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{copyToFSTs}
\alias{copyToFSTs}
\title{Convert DTF files to FST files}
\usage{
copyToFSTs(ListOfItems)
}
\arguments{
\item{ListOfItems}{List of named vectors. If ListOfItems is a named vector, then Source and Target locations. If ListOfItems is a list of dataframes than the locations are the names of the list elements.}
}
\value{
None. FTS files have been created on disk.
}
\description{
* NOT YET IMPLEMENTED *
From a vector of DBF file locations (or a list of named (location of each DBF data file) with the value being a data.frame), create an FTS file.
}
