% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{createNLN}
\alias{createNLN}
\title{Create a R CRAN package TTR function without the Leading NA error}
\usage{
createNLN(x, f, ...)
}
\arguments{
\item{x}{Default none. Required. Single column xts data.}

\item{f}{Default none. Required. TTR function.}

\item{...}{Dots passed to f.}
}
\value{
modified x
}
\description{
Non-Leading NAs (NLN).
Work-around of R CRAN package TTR functions:
Error in runSum(x, n) : Series contains non-leading NAs.
}
\examples{
\dontrun{
x <- xts::xts(c(NA,1,2,4,NA), zoo::as.Date(0:4))
SMA <- createNLN(x, f = TTR::SMA)
SMA(x, n = 2)
V1nln.SMA.n.2
1970-01-01            NA
1970-01-02            NA
1970-01-03           1.5
1970-01-04           3.0
1970-01-05           0.0
}
}
