% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dateAAIISIPro}
\alias{dateAAIISIPro}
\title{Get the Data Version of the AAII StockInvestor Pro}
\usage{
dateAAIISIPro(From = "C:/Program Files (x86)/Stock Investor/Professional")
}
\arguments{
\item{From}{String. Directory of AAII StockInvestor Pro installation directory.  Defaults to "C:/Program Files (x86)/Stock Investor/Professional".}
}
\value{
String. Date of the SIPro update, in days since the UNIX epoch (birthday of UNIX: January 1st, 1970). Returned is the "Current as of date" of StockInvestor Pro.  This is the same data found by doing Help -> About (and then reading the bottom line).
}
\description{
The American Association of Individual Investors has a software product
called StockInvestor Pro. Historically periodically e.g. monthly, but recently
daily, new updated corporate data of of 6000+ companies is updated and
re-available. The software includes (much) information from the following:
income statements, cash flow statements, balance sheets, stock prices,
analyst ratings, and company general information.
This function reads the file  Setup.dbf file. It is a dBASE file.
Note, the .dbf file can be opened interactively (and read)
using/through LibreOffice Calc, R package foreign function read.dbf, and
other applications that use the file shapefil.h written by Frank Warmerdam.
}
\examples{
\dontrun{
dateAAIISIPro()
[1] "18565"

zoo::as.Date(as.integer(dateAAIISIPro()))
[1] "2020-10-30"
}
}
\references{
\cite{foreign/src/shapefil.h
\url{https://github.com/cran/foreign/blob/master/src/shapefil.h}
}

\cite{Index of /shapelib/
\url{http://download.osgeo.org/shapelib/}
}
}
