% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbAddKeyEM}
\alias{dbAddKeyEM}
\title{Add Key or Check Constraint}
\usage{
dbAddKeyEM(
  connName,
  name,
  colname,
  if.not.exists = FALSE,
  only = FALSE,
  const.name = characater(),
  type = c("primary", "foreign", "check"),
  check.by = character(),
  reference,
  colref,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{name}{A character string, or a character vector, specifying a PostgreSQL table name.}

\item{colname}{A character string specifying the name of the column to which the key will be assign; alternatively, a character vector specifying the name of the columns for keys spanning more than one column.}

\item{if.not.exists}{Logical. Default is FALSE. If TRUE, use pg_ tables to check that in the namespace(schema) the constraint name does not exist.  Needed for the check to work is that parameter "const.name" must also be provided.}

\item{only}{Logical. Default is FALSE. Whether to add to apply this key just to this parent table(TRUE). Otherwise, also apply this constraint to inherited tables(FALSE).}

\item{const.name}{String. Name of the constraint.}

\item{type}{The type of the key, either "primary" or "foreign" or "check" constraint}

\item{check.by}{If type = "check", then the value of the "check".  Ignored otherwise.}

\item{reference}{A character string specifying a foreign table name to which the foreign key will be associated (ignored if type == "primary").}

\item{colref}{A character string specifying the name of the primary key in the foreign table to which the foreign key will be associated; alternatively, a character vector specifying the name of the columns for keys spanning more than one column (ignored if type == "primary").}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
TRUE if the key was successfully added.
}
\description{
Add a primary or foreign key or check constraint to a table column.
}
