% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbAttachPartEM}
\alias{dbAttachPartEM}
\title{Add or Remove a Partition}
\usage{
dbAttachPartEM(
  conn,
  dbobject = "table",
  name,
  partition,
  part.bound.value,
  action = c("attach", "drop"),
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A connection object.}

\item{dbobject}{A character string specifying if the the type of the parent database object.  This can be "table"(default) or "index".}

\item{name}{A character string specifying a PostgreSQL already partitioned table or index name. Note, a table can not be converted to a partitioned table and vice-versa.  ALTER has no option.}

\item{partition}{A character string specifying the name of the partition.}

\item{part.bound.value}{Of the attaching the partition, the partitions partition bound single value.}

\item{action}{A character string specifying if the column is to be added ("attach", default) or removed ("drop").}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
TRUE if the partition was successfully added or removed.
}
\description{
From a partitioned table or partitioned index add or remove a partition.
}
