% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbColumnEM}
\alias{dbColumnEM}
\title{Add or Remove a Column}
\usage{
dbColumnEM(
  conn,
  name,
  colname,
  action = c("add", "drop"),
  coltype = "integer",
  cascade = FALSE,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A connection object.}

\item{name}{A character string specifying a PostgreSQL table name.}

\item{colname}{A character string specifying the name of the column}

\item{action}{A character string specifying if the column is to be added ("add", default) or removed ("drop").}

\item{coltype}{A character string indicating the type of the column, if action = "add".}

\item{cascade}{Logical. Whether to drop foreign key constraints of other tables, if action = "drop".}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
TRUE if the column was successfully added or removed.
}
\description{
Add or remove a column to/from a table.
}
