% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateDbaseEM}
\alias{dbCreateDbaseEM}
\title{Database Creation}
\usage{
dbCreateDbaseEM(
  connName,
  dbname,
  owner = dbname,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{dbname}{String. Required.  Potential database in the cluster.}

\item{owner}{String. Database owner. Defaults to parameter "dbname".}

\item{env}{Environment.  Default is the .Global environment.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a a database in the cluster.
}
\examples{
\dontrun{
dbCreateDbaseEM(dbname = "r_user")
}
}
