% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExecuteEM}
\alias{dbExecuteEM}
\title{Executes in the  database}
\usage{
dbExecuteEM(
  connName,
  Statement,
  time_zone = "UTC",
  client_encoding = "UTF8",
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{Statement}{String. Required.  DML/DDL/DCL to execute}

\item{time_zone}{Execution time zone. Default is "UTC".}

\item{client_encoding}{Execution encoding. Default is "UTF8".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
Execution of "conn" from the database or disconnects "connName" from the database and removes it from the environment "env".
}
\description{
Disconnects the R object passed "conn".
Alternately, can disconnect and remove the DBI connection object connName (default is "connEM") from the environment "env".
The function will just do one or the other, but not both.  Does NOT remove the passed R object "conn".
}
\examples{
\dontrun{
dbExecuteEM(Statement = "CREATE TABLE xyz();")
}
}
