% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExistsDbaseEM}
\alias{dbExistsDbaseEM}
\title{Database Existence?}
\usage{
dbExistsDbaseEM(conn, dbname, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{dbname}{String. Required.  Potential database in the cluster.}

\item{...}{Dots passed.}
}
\value{
TRUE(exists) or FALSE(not exists)
}
\description{
Determine if a database exists in the cluster.
}
\examples{
\dontrun{
 dbExistsDbaseEM(conn, dbname = "r_user_econmodel")
 dbExistsDbaseEM(conn, dbname = "rtmp")
}
}
