% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetCurrentuserEM}
\alias{dbGetCurrentuserEM}
\title{PostgreSQL Current User}
\usage{
dbGetCurrentuserEM(conn, exec = TRUE, display = TRUE)
}
\arguments{
\item{conn}{PostgreSQL DBI connection.}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}
}
\value{
Current user
}
\description{
Get PostgreSQL CURRENT_USER.
}
\examples{
\dontrun{
dbGetCurrentuserEM(get("connEM"))
}
}
