% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetInfoExtraEM}
\alias{dbGetInfoExtraEM}
\title{Extra Information About The RPostgreSQL Connection}
\usage{
dbGetInfoExtraEM(conn, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{PostgreSQL DBI connection}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
R list of Strings of properties.
}
\description{
Adds extra information: current_schema, search_path, temp_dbname, econmodel_db_dbname, client_encoding, and time_zone.
It will not report the working database (econmodel_db_dbname)  (if it does not yet exist).
It will not report the user temporary database (temp_dbname) (if it does not yet exist).
}
\examples{
\dontrun{
dbGetInfoExtraEM(conn)
dbGetInfoExtraEM(get("connEM"))
}
}
