% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbIndexEM}
\alias{dbIndexEM}
\title{Create an index}
\usage{
dbIndexEM(
  connName,
  name,
  colname,
  if.not.exists = FALSE,
  only = FALSE,
  idxname,
  unique = FALSE,
  method = c("btree", "hash", "rtree", "gist"),
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{name}{A character string specifying a PostgreSQL table name.}

\item{colname}{A character string, or a character vector specifying the name of the column to which the key will be associated; alternatively, a character vector specifying the name of the columns to build the index.}

\item{if.not.exists}{Logical. Default is FALSE. If TRUE, add the " IF NOT EXISTS " clause.}

\item{only}{Logical. Default is FALSE. Whether to add to apply this key just to this parent table(TRUE). Otherwise, also apply this index to inherited indexes(FALSE).}

\item{idxname}{A character string specifying the name of the index to be created. By default, this uses the name of the table (without the schema) and the name of the columns as follows: <table_name>_<column_names>_idx.}

\item{unique}{Logical. Causes the system to check for duplicate values in the table when the index is created (if data already exist) and each time data is added. Attempts to insert or update data which would result in duplicate entries will generate an error.}

\item{method}{The name of the method to be used for the index. Choices are "btree", "hash", "rtree", and "gist". The default method is "btree", although "gist" should be the index of choice for PostGIS spatial types (geometry, geography, raster).}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
TRUE if the index was successfully created.
}
\description{
Defines a new index on a PostgreSQL table.
}
