% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbIsConnectedEM}
\alias{dbIsConnectedEM}
\title{Test The RPostgreSQL Connection}
\usage{
dbIsConnectedEM(connName, env, display = TRUE, exec = TRUE)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
TRUE(connected) or FALSE(otherwise)
}
\description{
Is the connection not expired or not valid?
}
\examples{
\dontrun{
dbIsConnectedEM()
}
}
