% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbListInheritEM}
\alias{dbListInheritEM}
\title{Of a PostgreSQL Object, Get Inherited or Inherit_From Objects}
\usage{
dbListInheritEM(
  connName,
  name,
  side = "parent",
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{name}{String. Required. Name of the object.}

\item{side}{String. View from the "parent side. Default is "parent".  Alternately, the view from the "child side" is the value "child".}

\item{env}{Environment.  Default is the .Global environment.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
Dataframe of inherited or inherit_from "objects" (if any) (with schema name).
}
\description{
Of the PostgreSQL object, get the child/parent objects (if any) with the schema.
}
\details{
Of column results named *_RELKIND, see the table pg_namespace column relkind.
These values are the following: r = ordinary table, i = index, S = sequence, t = TOAST table, v = view, m = materialized view, c = composite type, f = foreign table, p = partitioned table, I = partitioned index.

Of column results named like '_PART_KEY_DEF' and '_PART_BOUND', these are the partition object key (LISTs, RANGEs, and HASHs) and the partitions (child objects) bounds (FOR VALUES . . .) (of the partition object).  See the references.
}
\examples{
\dontrun{
dbListInheritEM(name = "sample")
}
}
\references{
\cite{Get number of partitions in PostgreSQL database
\url{https://stackoverflow.com/questions/8634144/get-number-of-partitions-in-postgresql-database }
}

\cite{How to identify the ranges over which a postgres table was partitioned?
\url{https://dba.stackexchange.com/questions/221277/how-to-identify-the-ranges-over-which-a-postgres-table-was-partitioned}
}
}
