% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{dbListInheritsEM}
\alias{dbListInheritsEM}
\title{Of a PostgreSQLTable Get Inherited Tables}
\usage{
dbListInheritsEM(conn, name, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{PostgreSQLConnection . Required.}

\item{name}{String. Required. Name of the table.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
Dataframe of inherited "child tables" (if any) (with schema name).
}
\description{
Of the PostgreSQL table get the child tables" (if any) (with schema name).
}
\examples{
\dontrun{
dbListInheritsEM(conn, name = "sample")
}
}
\references{
\cite {Get number of partitions in PostgreSQL database
url{ https://stackoverflow.com/questions/8634144/get-number-of-partitions-in-postgresql-database }
}
}
