% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{dbListPartitionsEM}
\alias{dbListPartitionsEM}
\title{Of a PostgreSQL Partioned Table Get Partitions}
\usage{
dbListPartitionsEM(conn, name, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{PostgreSQLConnection . Required.}

\item{name}{String. Required. Name of the table.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
Dataframe of information of "partition child tables" (if any) (with schema name).
}
\description{
Of the PostgreSQL (r)egular table or (p)artitioned table get the information of "partition child tables" (if any) (with schema name).
}
\examples{
\dontrun{
dbListPartitionsEM(conn, name = "sample")
}
}
