% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{dbListPartMetaEM}
\alias{dbListPartMetaEM}
\title{Of a PostgreSQL Table Get Partition Metadata}
\usage{
dbListPartMetaEM(conn, name, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{PostgreSQLConnection . Required.}

\item{name}{String. Required. Name of the table.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
Dataframe of information of "partition key definition" (if any) and/or "partition bound" (if any).
}
\description{
Of the PostgreSQL (r)egular table or (p)artitioned table get the "partition key definition" (if any) and/or "partition bound" (if any).
}
\examples{
\dontrun{
dbGetPartMetaEM(conn, name = "sample")
}
}
