% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbLoginEM}
\alias{dbLoginEM}
\title{Connect to the econModel database.}
\usage{
dbLoginEM(
  driver,
  connName,
  user,
  password = user,
  host,
  dbname = user,
  port,
  tty,
  options,
  forceISOdate,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{driver}{String. Defaults to getOption("econmodel_db_driver"). String.  Default is "PostgreSQL".  Currently only an implementation exists using PostgreSQL and PostgreSQL-like databases.}

\item{connName}{String.  Name of the database connection object.  The default is "connEM".}

\item{user}{String. Defaults to getOption("econmodel_db_user").}

\item{password}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_password").}

\item{host}{String. Defaults to getOption("econmodel_db_host").}

\item{dbname}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_dbname").}

\item{port}{Integer. Defaults to getOption("econmodel_db_port").}

\item{tty}{Default to getOption("econmodel_db_tty").}

\item{options}{Defaults to getOption("econmodel_db_dboptions").}

\item{forceISOdate}{Logical. Default is getOption("econmodel_db_forceISOdate").}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
Invisibly a DBI connection in an object named "connEM" is created, connected, and assigned to the environment "env".
}
\description{
Currently is only implemented to work on PostgreSQL or PostgreSQL-like databases.
Try to connect or try to connect as "user".
}
\examples{
\dontrun{
# by default, this the lowercase name of the tempdir()
options(econmodel_db_storage_name = "postgres")
getOption("econmodel_db_user")
# In rstudio to a Restart of R
devtools::load_all(".")
getOption("econmodel_db_user")
[1] "postgres"
dbLoginEM()
# User login succeeded
}
}
