% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbLogoutEM}
\alias{dbLogoutEM}
\title{Disconnect from the econModel database}
\usage{
dbLogoutEM(connName, env, ...)
}
\arguments{
\item{connName}{String.  Name of the database connection object. Optional. This may be "passed" instead of  "conn".}

\item{env}{Environment. Default is the global environment .GlobalEnv.  Location of the connection object "connName".}

\item{...}{Dots passed.}
}
\value{
Disconnects "connEM" from the database or disconnects "connName" from the database and removes it from the environment "env".
}
\description{
Disconnect from the econModel database
}
\examples{
\dontrun{
dbLogoutEM() # default is connection variable "connEM" in the .GlobalEnv
}
}
