% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbObjectNameFix}
\alias{dbObjectNameFix}
\title{Format input for database schema/object names.}
\usage{
dbObjectNameFix(
  conn = NULL,
  o.nm,
  as.identifier = TRUE,
  dbQuote = "Identifier",
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A connection object. Must be provided but can be set NULL,
where a dummy connection will be used.}

\item{o.nm}{Object name string, length 1-2.}

\item{as.identifier}{Boolean whether to return (schema,table) name as database
sanitized identifiers (TRUE) or as regular character (FALSE)}

\item{dbQuote}{String. Only used when "as.identifier = TRUE". Default is "Identifier". Alternately, this value can be "Literal."}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
character vector of length 2. Each character element is in
    (escaped) double-quotes when as.identifier = TRUE.
}
\description{
This is a near copy of the R CRAN package rpostgis function dbTableNameFix.
}
\details{
Internal rpostgis function to return common (length = 2) schema
and table name vector from various table and schema + table name
inputs.
}
\examples{
\dontrun{
name <- c("schema","table")
dbObjectNameFix(conn, name)

#current search path schema is added to single-length character object (if only table is given)
name<-"table"
dbObjectNameFix(conn,name)

#schema or table names with double quotes should be given exactly as they are
(make sure to wrap in single quotes in R):
name <- c('sch"ema','"table"')
dbObjectNameFix(conn,name)
}
}
\keyword{internal}
