% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbSetPerformanceEM}
\alias{dbSetPerformanceEM}
\title{PostgreSQL Performance}
\usage{
dbSetPerformanceEM(conn, exec = TRUE, display = TRUE)
}
\arguments{
\item{conn}{PostgreSQL DBI connection.}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}
}
\value{
PostgreSQL parameters are set
}
\description{
Set PostgreSQL memory parameters.
}
\examples{
\dontrun{
dbSetPerformanceEM(get("connEM"))
}
}
