% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbWriteTableEM}
\alias{dbWriteTableEM}
\title{Results column names are always in UPPERCASE.}
\usage{
dbWriteTableEM(
  conn,
  DfName = substitute(Df),
  Df,
  FillNull = TRUE,
  AddID = FALSE,
  RowNames = FALSE,
  PgUpdateSeq = FALSE,
  lowerDfName = TRUE,
  lowerColNames = TRUE,
  replaceDotUsingUnderscore = TRUE,
  ...
)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by dbConnect().}

\item{DfName}{String.  Default is substitute(Df). The name of the table to which the data frame is to be loaded.}

\item{Df, }{data.frame. Required. To be loaded to the database.}

\item{FillNull}{Logical.  Default is TRUE. Should new db present fields be added to the data.frame before it is loaded?}

\item{AddID}{Logical. Default is FALSE.  Should a new column should be added for the database id?}

\item{RowNames}{Logical. Default is FALSE.  Should the row names be loaded as a separate column? (unlike the original dbWriteTable, default is FALSE).}

\item{PgUpdateSeq}{Logical. Default is FALSE. Should the table primary key's sequence be updated to the highest id value +1? (Postgres specific)}

\item{lowerDfName}{Logical. Default is TRUE. Make the target database table name to be in lowercase.}

\item{lowerColNames}{Logical. Default is TRUE. Make the target database table column names to be in lowercase.}

\item{replaceDotUsingUnderscore}{Logical. Default is TRUE. Make the target database table column names internal "dots" be converted to underscores(_).}

\item{...}{Dots. Other parameters passed to R CRAN package DBI dbWriteTable.}
}
\description{
Thin wrapper over R CRAN package caroline function dbWriteTable2
}
