% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCombine.R
\name{DetectFullRows}
\alias{DetectFullRows}
\title{detect rows with non-NA vaues}
\usage{
DetectFullRows(x, Var = colnames(x), Ele = "All")
}
\arguments{
\item{x}{data.frame}

\item{Var}{columns determining complete cases.
Can be column names or positions.
Var is the determiner of full rows.
Other columns are ignored.}

\item{Ele}{"All"(default), of elements in Var, the test results.
The other option is "Any".}
}
\value{
df with rows of all-NA removed
}
\description{
\preformatted{

complete.cases does not work on tibbles!
This function is written in package DataCombine style.

}
}
\examples{
\dontrun{
data.frame(A = c(1,NA,NA), B = c(11,12,NA), C = c(101,102,103))
   A  B   C
1  1 11 101
2 NA 12 102
3 NA NA 103

DetectFullRows(data.frame(A = c(1,NA,NA), B = c(11,12,NA), C = c(101,102,103)),
  Var = c("A","B"))
[1]  TRUE FALSE FALSE
}
}
