% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCombine.R
\name{DetectOnlyNonEmptyVarsInRows}
\alias{DetectOnlyNonEmptyVarsInRows}
\title{detect rows with columns of non-NA vaues and other columns of non-NA values}
\usage{
DetectOnlyNonEmptyVarsInRows(x, Var = colnames(x), Ele = "All", EleO = "All")
}
\arguments{
\item{x}{data.frame}

\item{Var}{columns determining complete cases.
Can be column names or positions.}

\item{Ele}{"All"(default), of elements in Var, the test results.
Other option is "Any".}

\item{EleO}{"All"(default), of elements in 'other than Var', the test results.
The other option is "Any".}
}
\value{
df with rows of all-NA removed
}
\description{
\preformatted{

Detect columns(Var) where the values are
not-NA (determined by Ele) but
the rest of the columns are
NA everywhere else (determined by EleO).

The function is written in package DataCombine style.

}
}
\examples{
\dontrun{
res <- data.frame(A = c(1,NA,NA,NA,5,6), B = c(11,12,NA,NA,15,NA), C = c(101,102,103,NA,NA,NA))
res
   A  B   C
1  1 11 101
2 NA 12 102
3 NA NA 103
4 NA NA  NA
5  5 15  NA
6  6 NA  NA

DetectOnlyNonEmptyVarsInRows(res, Var = c("A","B"))
[1] FALSE FALSE FALSE FALSE  TRUE FALSE
}
}
