% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dfServerFieldsClassesEM}
\alias{dfServerFieldsClassesEM}
\title{Given an R data.frame, Get The Equivalent Database Server Column Types
If a data.frame would be promoted to be on a database server, then these would be the equivalent Server datatypes.}
\usage{
dfServerFieldsClassesEM(connName, value, env, exec = TRUE, display = TRUE)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{value}{data.frame}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}
}
\value{
data.frame. Variables are "name" and "type"
}
\description{
Given an R data.frame, Get The Equivalent Database Server Column Types
If a data.frame would be promoted to be on a database server, then these would be the equivalent Server datatypes.
}
\details{
\preformatted{

typical SQLite results of TYPE values:
"TEXT"  "REAL" "INTEGER"

typical PostgreSQL results of TYPE values:
"text"  "float8" "bool"  "integer" "date"
}
}
