% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{diffXts}
\alias{diffXts}
\title{Lag and/or Difference/Use(rerun) a function(Fun) Upon its xts Object}
\usage{
diffXts(
  x,
  lag = 1,
  differences = 1,
  arithmetic = TRUE,
  log = FALSE,
  na.pad = TRUE,
  Fun = diff,
  ...
)
}
\arguments{
\item{x}{as diff.xts}

\item{lag}{as diff.xts}

\item{differences}{as diff.xts.
Also, knows as "number of (re)runs - 1" of function(Fun).
Runs use as its input data, the results data from the the previous run (if any).s}

\item{arithmetic}{as diff.xts}

\item{log}{as diff.xts}

\item{na.pad}{as diff.xts}

\item{Fun}{difference-ing function.
Meant to change the xTs in some way.
(Default diff (expected: xts::diff.xts)).
Should accept or (accept and ignore) the parameters: lag;
for S3 compatibility, differences; for xts compatiblity,
arithmetic, log, and/or na.pad.}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Based on R CRAN package xts function diff.xts
}
}
\examples{
\dontrun{

# diffXts examples

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

diffXts(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
        V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       -3        8
1970-01-03       -2       16

# same as above
diffXts(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)), Fun = AC)
           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       -3        8
1970-01-03       -2       16

diffXts(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)),
       differences = 2, Fun = AC)

           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       NA       NA
1970-01-03        1        8

diffXts(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)), Fun = RC)
           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       -2        2
1970-01-03       -1        2

diffXts(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)),
        differences = 2, Fun = RC)

           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       NA       NA
1970-01-03      0.5        1

}
}
