% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{disconnectEM}
\alias{disconnectEM}
\title{Disconnect from and econModel database}
\usage{
disconnectEM(connName, env)
}
\arguments{
\item{connName}{String.  Required.  Name of the database connection object}

\item{env}{Environment.  Default is the global environment .GlobalEnv.  Location of the connection object "connName"}
}
\value{
Disconnects "connName" and removes it from the environment "env".
}
\description{
Also, removes the DBI connection object from the environment "env".
}
