% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModel.R
\docType{package}
\name{econModel}
\alias{econModel}
\alias{econModel-package}
\title{econModel: Meaningful Social and Economic Data}
\description{
Collect and display meaningful social and economic data
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/AndreMikulec/econModel/issues}
}

}
\author{
\strong{Maintainer}: Andre Mikulec \email{Andre_Mikulec@Hotmail.com}

}
