% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{estimLastUpdated}
\alias{estimLastUpdated}
\title{Determine LastUpdated(Publishing) Dates}
\usage{
estimLastUpdated(
  x,
  Calendar = "UnitedStates/GovernmentBond",
  Frequency,
  LastUpdated = NULL,
  NdayInMonth = NULL,
  LastOfDateRange = NULL
)
}
\arguments{
\item{x}{Date times of observation dates}

\item{Calendar}{Default is "UnitedStates/GovernmentBond". Calendar to use.  See ?? RQuantLib::Calendars}

\item{Frequency}{No default (required). Values can be Daily(not implemented yet), Weekly(not implemented yet), Monthly, and Quarterly}

\item{LastUpdated}{Date time.  Default NULL.  Date time of the published date of the newest(latest) observation}

\item{NdayInMonth}{Integer.  Default NULL.  This only applied to the Frequency of "Quarterly" or "Monthly". If this parameter is present, then this parameter provides additiona information to parameter LastUpdated.  Use this parameter NdayInMonth to as the Day after the Reference period to estimate.  See ? timeDate::timeNthNdayInMonth}

\item{LastOfDateRange}{Default NULL.  If not present, this value will be taken from the last observation of x.  This parameter value represents the earliest of the time between an observation and its corresponding LastUpdated Date time.}
}
\value{
vector of Date times
}
\description{
Estimate the time distance between observations Dates/times and their corresponding LastUpdated(Publishing) Dates/times.
}
\examples{
\dontrun{
Title:               Gross Domestic Product
Series ID:           GDP
Source:              U.S. Bureau of Economic Analysis
Release:             Gross Domestic Product
Seasonal Adjustment: Seasonally Adjusted Annual Rate
Frequency:           Quarterly
Units:               Billions of Dollars
Date Range:          1947-01-01 to 2020-07-01
Last Updated:        2020-10-29 7:52 AM CDT

atr <- fredAttributes("GDP")

library(quantmod)

getSymbols("GDP", src = "FRED")

estimLastUpdated(index(getSymbols("GDP")),
  Frequency = atr$Frequency,
  LastUpdated = atr$LastUpdated,
  LastOfDateRange = tail(strsplit(atr$DateRange, " to "),1)
)

Title:               Unemployment Rate
Series ID:           UNRATE
Source:              U.S. Bureau of Labor Statistics
Release:             Employment Situation
Seasonal Adjustment: Seasonally Adjusted
Frequency:           Monthly
Units:               Percent
Date Range:          1948-01-01 to 2020-10-01
Last Updated:        2020-11-06 7:47 AM CST

# LastUpdated (updated) during the first Friday of every month
# if that "first Friday" is not available, then the
# the updated occurs at the next Friday
#
# Schedule of Releases for the Employment Situation
https://www.bls.gov/schedule/news_release/empsit.htm

# PAY & LEAVE (UNRATE data is not released during these days)
https://www.opm.gov/policy-data-oversight/pay-leave/federal-holidays/#url=2020

atr <- fredAttributes("UNRATE")

library(quantmod)

getSymbols("UNRATE", src = "FRED")

# UNRATE is released on the First Friday that is
# not a weekend day no holiday
#
tail(estimLastUpdated(index(UNRATE),
  Frequency =  head(strsplit(atr$Frequency, ", ")[[1]],1),
  LastUpdated = atr$LastUpdated,
  LastOfDateRange = tail(strsplit(atr$DateRange, " to ")[[1]],1)
),4)

[1] "2020-08-06" "2020-09-08" "2020-10-07" "2020-11-05"

# first Friday (better)
tail(estimLastUpdated(index(UNRATE),
                 Frequency =  head(strsplit(atr$Frequency, ", ")[[1]],1),
                 LastUpdated = atr$LastUpdated,
                 NdayInMonth = 5L, # 1st Friday
                 LastOfDateRange = tail(strsplit(atr$DateRange, " to ")[[1]],1)
),4)

[1] "2020-08-07" "2020-09-04" "2020-10-02" "2020-11-06"

}
}
\references{
\cite{Schedule of Releases for the Employment Situation
\url{https://www.bls.gov/schedule/news_release/empsit.htm}}
}
