% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{fileAndExt}
\alias{fileAndExt}
\title{FileName With Its Extension}
\usage{
fileAndExt(SubPathFileExt, Separ = "/", PathReq = TRUE, ...)
}
\arguments{
\item{SubPathFileExt}{String. Path to the file including FileName and its (after the dot ".") extension.}

\item{Separ}{String. Regular expression separating the FileName with its extension from the rest of the path.}

\item{PathReq}{Logical. Send SubPathFileExt through normalizePath.}

\item{...}{Dots passed.}
}
\value{
String.
}
\description{
Get the FileName with its extension.
}
\examples{
\dontrun{
fileAndExt(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
))
}
}
