% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{fileName}
\alias{fileName}
\title{FileName Without Its Extension}
\usage{
fileName(SubPathFileExt, Separ = "[.]", PathReq = TRUE, ...)
}
\arguments{
\item{SubPathFileExt}{String. Path to the file including FileName and its (after the dot ".") extension.}

\item{Separ}{String. Regular expression separating the FileName from its extension.}

\item{PathReq}{Logical. Send SubPathFileExt through normalizePath.}

\item{...}{Dots passed.}
}
\value{
String.
}
\description{
Get the FileName without its extension.
}
\examples{
\dontrun{
fileName(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
))
}
}
