% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{filePartName}
\alias{filePartName}
\title{FilePartName}
\usage{
filePartName(SubPathFileExt, Separ = "[.]", PathReq = TRUE, Position, ...)
}
\arguments{
\item{SubPathFileExt}{String. Path to the file including FileName and its (after the dot ".") extension.}

\item{Separ}{String. Regular expression separating the FileName from its extension.}

\item{PathReq}{Logical. Send SubPathFileExt through normalizePath.}

\item{...}{Dots passed.}
}
\value{
String.
}
\description{
Get the FilePartName.
}
\examples{
\dontrun{
filePartName(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
), Position = "first")
}
}
