% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{filePositionName}
\alias{filePositionName}
\title{FileName FilePositionName Without Its Extension}
\usage{
filePositionName(SubPathFileExt, Separ = "_", PathReq = TRUE, Position, ...)
}
\arguments{
\item{SubPathFileExt}{String. Path to the file including FileName and its (after the dot ".") extension.}

\item{Separ}{String. Regular expression separating the FilePostionName from its FileName.}

\item{PathReq}{Logical. Send SubPathFileExt through normalizePath.}

\item{Position}{String or Number. Required. Can be "first", 1, "last", or a number of the position.}

\item{...}{Dots passed.}
}
\value{
String.
}
\description{
Get the FileName FilePositionName without its extension.
}
\examples{
\dontrun{
filePositionName(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
), Position = "first")
}
}
