% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{filePostFixName}
\alias{filePostFixName}
\title{FileName FilePostFixName Without Its Extension}
\usage{
filePostFixName(SubPathFileExt, Separ = "_", PathReq = TRUE, ...)
}
\arguments{
\item{SubPathFileExt}{String. Path to the file including FileName and its (after the dot ".") extension.}

\item{Separ}{String. Regular expression separating the FilePostFixName from its FileName.}

\item{PathReq}{Logical. Send SubPathFileExt through normalizePath.}

\item{...}{Dots passed.}
}
\value{
String.
}
\description{
Get the FileName FilePostFixName without its extension.
}
\examples{
\dontrun{
filePostFixName(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
))
}
}
