% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{first}
\alias{first}
\title{Return First n Elements of A Data Object}
\usage{
first(x, ...)
}
\arguments{
\item{x}{1 or 2 dimensional data object.}

\item{...}{Dots passed.}
}
\value{
first elements or zero length element
}
\description{
A generic function to return the first or first elements or rows of a vector or two-dimensional data object.
This is a wrapper over the R CRAN package xts S3 function first.
Different here is that if n elements can not be returned, then instead, zero elements are returned.
}
