% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{firstly}
\alias{firstly}
\title{Tries to Return First n Elements of A Data Object}
\usage{
firstly(x, ...)
}
\arguments{
\item{x}{1 or 2 dimensional data object.}

\item{...}{Dots passed. Typically, tries to gather up to "n" elements to return.}
}
\value{
first elements or zero length element
}
\description{
A generic function to return the first or first elements or rows of a vector or two-dimensional data object.
This is a wrapper over the R CRAN package xts S3 function first.

It tries to return (if available) up to "n" elements
}
