% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModelSupport.R
\name{forceAssignInNamespace}
\alias{forceAssignInNamespace}
\title{Force the assignment of new items in a namespace}
\usage{
forceAssignInNamespace(x, value, namespace)
}
\arguments{
\item{x}{name of the symbol/function inside a string}

\item{value}{new value of the symbol/function}

\item{namespace}{namespace}
}
\description{
Adapted from Rcpp. Based on the original function forceAssignMyNamespace by Willem Ligtenberg.
}
\examples{
\dontrun{
library(econModel)
forceAssignMyNamespace("prnt", function(x) {
  print(x)
}, namespace = "econModel")

econModel::prnt("Do it.")
}
}
\references{
\cite{forceAssignMyNamespace
\url{https://github.com/openanalytics/Rango/blob/adc99e077b71c8c6826cabb7ff1266050898718a/Rango/R/utils.R}
}
}
\author{
Willem Ligtenberg
}
