% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{fredAttributes}
\alias{fredAttributes}
\title{Download attributes of Times Series of Federal Reserve Economic Data - FRED(R)}
\usage{
fredAttributes(Symbol)
}
\arguments{
\item{Symbol}{specifying the name of the symbol to be queried}
}
\value{
data.frame of attributes
}
\description{
Data typically includes the Title, Series ID, Source, Release, Seasonal Adjustment, Frequency, Units, Date Range, Last Updated, and Notes.
}
\examples{
\dontrun{

# Smoothed U.S. Recession Probabilities
# Source: Piger, Jeremy Max, Chauvet, Marcelle
# https://fred.stlouisfed.org/data/RECPROUSM156N.txt
#
print(str(fredAttributes("RECPROUSM156N")))

'data.frame':	1 obs. of  10 variables:
$ Title             : chr "Smoothed U.S. Recession Probabilities"
$ SeriesID          : chr "RECPROUSM156N"
$ Source            : chr "Piger, Jeremy Max, Chauvet, Marcelle"
$ Release           : chr "U.S. Recession Probabilities"
$ SeasonalAdjustment: chr "Not Seasonally Adjusted"
$ Frequency         : chr "Monthly"
$ Units             : chr "Percent"
$ DateRange         : chr "1967-06-01 to 2020-09-01"
$ LastUpdated       : chr "2020-11-02 7:01 AM CST"
$ Notes             : chr "Smoothed recession probabilities for the United"
}
}
