% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{getModelData}
\alias{getModelData}
\title{Update model's dataset}
\usage{
getModelData(x, na.rm = TRUE, source.envir = NULL, ...)
}
\arguments{
\item{x}{Default is none. Required. An object of class quantmod.}

\item{na.rm}{Logical. Defaults to TRUE. Remove NA values.}

\item{source.envir}{Default is NULL. First place to search for Symbols.}

\item{...}{Dots passed to getSymbolsEnv.}
}
\value{
Returns object of class quantmod.OHLC
}
\description{
getModelData updates the  currently specified or built model with most recent data.
getModelData is a re-implementation of the R CRAN package getModelData with added flexibility in source.envir and Dots.
}
\details{
\preformatted{
Primarily Used within specify model calls, getModelData is used
to retrieve the appropriate underlying variables, and
apply model specified transformations automatically.
It can be used to also update a current model in memory
with the most recent data.
}
}
\author{
Andre Mikulec (re-implementation)

Jeffrey Ryan
}
