% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{getSymbols.AAIISentiment}
\alias{getSymbols.AAIISentiment}
\title{American Association of Individual Investors (AAII) Weekly Sentiment Survey}
\usage{
getSymbols.AAIISentiment(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{A character vector specifying the names of each symbol to be loaded.
Possible Symbols are the following:
"AAIISentimentSurveyData" (means get all of the columns) xor
one specific column name. e.g. "Bullish", "Neutral", or "Bearish", (or Others).}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{Dots passed.}
}
\value{
A call to getSymbols.AAIISentiment will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
Data represents what direction members feel the
stock market will be in next 6 months.

The sentiment survey measures the percentage of individual investors who are
bullish, bearish, and neutral on the stock market short term;
individuals are polled from the AAII Web site on a weekly basis.
Only one vote per member is accepted in each weekly voting period.

The latest published data seems to be delivered every Friday.

Internally, this reads a (downloaded) Microsoft Excel spreadsheet
and collects just the one latest published item by scraping the
one new (the latest) observation from the AAII web site.
}
\examples{
\dontrun{
# common usage
getSymbols("BullBearSpread", src = "AAIISentiment")

# The dates are as of Thursday each week.
getSymbols(c("Bullish", "Neutral", "Bearish"), src = "AAIISentiment")

# all columns in one xts object
getSymbols("AAIISentimentSurveyData", src = "AAIISentiment")
}
}
\references{
\cite{AAII Investor Sentiment Survey
\url{https://www.aaii.com/sentimentsurvey}}

\cite{Sentiment Survey Past Results
\url{https://www.aaii.com/sentimentsurvey/sent_results}}

\cite{Charles Rotblut, Analyzing the AAII Sentiment Survey Without Hindsight, AAII Journal / June 2014
\url{https://www.aaii.com/journal/article/analyzing-the-aaii-sentiment-survey-without-hindsight}}
}
\author{
Andre Mikulec (adapted original code to work with AAII sentiment data)

Jeffrey A. Ryan
}
