% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{getSymbols.USFedPhilForecasts}
\alias{getSymbols.USFedPhilForecasts}
\title{Survey of Professional Forecasters data from the Philadelphia Federal Researve Board of Governors (FED)}
\usage{
getSymbols.USFedPhilForecasts(
  Symbols,
  env,
  return.class = "xts",
  Fun = mean,
  ...
)
}
\arguments{
\item{Symbols}{A character vector.  Required. Specifies the names of each symbol to be loaded
To get all of the data use: "USFedPhilForecastsData".  Possible individual
Symbols are listed in the Forecasters documentation: (1) Historical Data Files for the Survey
of Professional Forecasters (2) Documentation: PDF Provides information on
all variables, transformations, and files in the survey.
E.g. "UNEMP". Symbols will be returned as (xts) data.}

\item{env}{Environment. Default is the global environment. Where to create objects.}

\item{Fun}{Function.  Default is mean. This is the aggregating function.  NA data cells are automatically removed from the data before the (remaining) data is sent to the aggregating function.}

\item{...}{Dots passed to Fun.}

\item{return.class.}{String. Desired class of returned object.  Default is "xts".
The value can be "zoo", "data.frame", or "xts".}
}
\value{
A call to getSymbols.USFedPhilForecasts will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
About the data, file Structure: Column Header Nomenclature and Forecast Horizons: ("1") through ("6") are quarterly forecasts and ("A") and ("B") annual forecasts.

The number ("1") represents the forecast of the quarter prior to the quarter in
which the survey is conducted. The forecasters know the values of the variables for this
quarter at the time they submit their projections in comparison to official revisions(corrections) of the official data.
The forecasters are permitted to forecast a revision.

The number ("2") represents the forecast for the current quarter, defined as
the quarter in which the survey is conducted. The numbers ("3") through ("6")
represent the forecasts for the four quarters after the current quarter.
The letters ("A") and ("B") represent annual average forecasts
for the current year (the year in which the survey is conducted) and
the following year.

INDUSTRY means: financial service provider ("1"), non-financial service provider ("2"), or do not know the industry ("3").  INDUSTRY is not shown in the output data.

About PR*, probability forecasts, read the documentation. This is "very" complicated.

About the date/time index, deadlines for responses are at late in the second week of the middle month of each quarter.
This report is released to the public in the second week of of the middle month of each quarter.
True deadline and news release dates for surveys prior to 1990:Q2 are not known.

Therefore, for simplicity in this R function, the index date is estimated to be at the mid-point of the quarter.

The 1990Q2 survey was not taken in real time, because the Philadelphia Fed
had not yet taken over the survey. Forecasters were asked to provide dated
forecasts from May 1990.
The 1996Q1 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.

The 2013Q4 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.
}
\note{
\preformatted{

Some, compare and contrasts, may be the
forecasters prediction v.s. FRED data of  what actually happened.

# unemployment

Civilian Unemployment Rate (UNEMP)
https://www.philadelphiafed.org/research-and-data/real-time-center
       /survey-of-professional-forecasters/data-files/unemp

Civilian Unemployment Rate (UNRATE)
https://fred.stlouisfed.org/series/UNRATE/

# prices ( e.g. inflation )

Price Index for Gross National Product/Gross Domestic Product (PGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center
       /survey-of-professional-forecasters/data-files/pgdp

Gross National Product (chain-type price index) (A001RV1Q225SBEA)
https://fred.stlouisfed.org/series/A001RV1Q225SBEA

# gross domestic product

Nominal Gross National Product/Gross Domestic Product (NGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center
       /survey-of-professional-forecasters/data-files/ngdp

Gross Domestic Product (GDP)
https://fred.stlouisfed.org/series/GDP

# corporate Profits

Corporate Profits After Tax (CPROF)
https://www.philadelphiafed.org/research-and-data/real-time-center
       /survey-of-professional-forecasters/data-files/cprof

Corporate Profits After Tax (without IVA and CCAdj) (CP)
https://fred.stlouisfed.org/series/CP
}
}
\examples{
\dontrun{
# common usage
# mean(the average) of the forecasts
getSymbols(c("UNEMP3", "INDPROD2"), src = "USFedPhilForecasts")

# standard deviation of forecasts
getSymbols("UNEMP3", src = "USFedPhilForecasts", Fun = sd, verbose = T)

# number of forecasters
getSymbols("UNEMP3", src = "USFedPhilForecasts", Fun = length)

# all columns in one VERY WIDE xts object
getSymbols("USFedPhilForecastsData", src = "USFedPhilForecasts")
}
}
\references{
\cite{Dates of deadlines for surveys from 1990:Q2 to the present (BROKEN LINK)
\url{https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-ofprofessional-forecasters/spf-release-dates.txt?la=en}
}

\cite{Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/surveys-and-data/real-time-data-research/survey-of-professional-forecasters}
}

\cite{Calendar of Events (Release(Publishing) Dates)
\url{https://www.philadelphiafed.org/calendar-of-events?release=survey-of-professional-forecasters#Economic-Release-Calendar}
}

\cite{Data Files - Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/surveys-and-data/data-files}
}

\cite{Surveys 1968:4–present, Individual Forecasts: Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/survey-of-professional-forecasters/historical-data/spfmicrodata.xlsx?la=en}
}

\cite{Caveats on Identification Numbers for Individual Forecasts, Individual Forecasts: Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/survey-of-professional-forecasters/spf-caveats.pdf?la=en}
}

\cite{Documentation, Individual Forecasts: Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/survey-of-professional-forecasters/spf-documentation.pdf?la=en}
}

\cite{Individual Forecasts: Survey of Professional Forecasters
\url{https://www.philadelphiafed.org/surveys-and-data/real-time-data-research/individual-forecasts}
}

\cite{Dean Croushore, Introducing The Survey of Professional Forecasters, Business review (Federal Reserve Bank of Philadelphia), 6(Nov):3-15
\url{https://www.researchgate.net/profile/Dean_Croushore/publication/5051610_Introducing_The_Survey_of_Professional_Forecasters/links/54369c5a0cf2dc341db3684b/Introducing-The-Survey-of-Professional-Forecasters.pdf}
}

\cite{Introducing: The Survey of Professional Forecasters
\url{https://www.researchgate.net/publication/5051610_Introducing_The_Survey_of_Professional_Forecasters}
}
}
\author{
Andre Mikulec
}
