% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{getVintages}
\alias{getVintages}
\title{Download vintage dates of Federal Reserve Economic Data - ALFRED(R)}
\usage{
getVintages(Symbol, src = "ALFRED")
}
\arguments{
\item{Symbol}{specifying the name of the symbol to be loaded}

\item{src}{see R CRAN package quantmod function getSymbols}
}
\value{
character vector of 'Last Updated' vintage dates
}
\description{
Downloads vintage dates of the published tail dates of 'Date Range' of the specified Symbol from 'research.stlouisfed.org'.
}
\details{
Used internally by the function getSymbols.ALFRED

Note, this function does not use the official API.  The official API usage can be performed using the R CRAN package alfred function get_alfred_series.  In some cases, this function getVintages may be faster.
}
\examples{
\dontrun{

library(quantmod)

# Smoothed U.S. Recession Probabilities
# Source: Piger, Jeremy Max, Chauvet, Marcelle
# https://fred.stlouisfed.org/data/RECPROUSM156N.txt
#
getVintages("RECPROUSM156N")
Read 2002 items

}
}
\references{
\cite{Download Data for Gross Domestic Product (GDP)
\url{https://alfred.stlouisfed.org/series/downloaddata?seid=GDP}
}

\cite{Download Data Help
\url{https://alfred.stlouisfed.org/help/downloaddata}
}
}
