% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{getYieldCurve}
\alias{getYieldCurve}
\title{Download US Treasury Yield Curve Data}
\usage{
getYieldCurve(
  base = "http://data.treasury.gov/feed.svc/DailyTreasuryYieldCurveRateData",
  year = NULL,
  month = NULL
)
}
\arguments{
\item{base}{String. Required. The base URL for the data service, defaulting to http://data.treasury.gov/feed.svc/DailyTreasuryYieldCurveRateData. If the month or year arguments are not NULL, then the function modifies this URL to parameterize the download request.}

\item{year}{The desired year number or NULL for all years (default)}

\item{month}{The desired month number or NULL for all months (default).}
}
\value{
xts object
}
\description{
Downloads US Treasury yield curve data from the US Treasury web site.
This is a fix and partial re-write of R CRAN package ustyc function getYieldCurve.

Omitted columns are like "^.*_DATE.*$" (because many character vector of many different date formats exist) and  like "^CUSIP_.*$" (and these are alphanumeric).
The omit-ion of columns most affects DailyTreasuryBillRateData.

Note, this is not "tuned for performance".  Tuning may be appropriate for a getSymbols.TreasuryYields. (Perhaps, this would be a future project.)
}
\examples{
\dontrun{

# https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=yield
xlist <- getYieldCurve("https://data.treasury.gov/feed.svc/DailyTreasuryYieldCurveRateData?$filter=month(NEW_DATE)\%20eq\%2012\%20and\%20year(NEW_DATE)\%20eq\%202020")
xlist
           BC_1MONTH BC_2MONTH BC_3MONTH BC_6MONTH BC_1YEAR BC_2YEAR BC_3YEAR BC_5YEAR BC_7YEAR
2020-12-01      0.07      0.07      0.09      0.10     0.12     0.17     0.22     0.42     0.68
2020-12-02      0.07      0.08      0.09      0.10     0.11     0.16     0.22     0.42     0.69
2020-12-03      0.08      0.08      0.08      0.09     0.10     0.16     0.21     0.40     0.67
2020-12-04      0.07      0.07      0.09      0.10     0.11     0.16     0.21     0.42     0.70
           BC_10YEAR BC_20YEAR BC_30YEAR BC_30YEARDISPLAY
2020-12-01      0.92      1.46      1.66             1.66
2020-12-02      0.95      1.50      1.70             1.70
2020-12-03      0.92      1.46      1.67             1.67
2020-12-04      0.97      1.53      1.73             1.73


# https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=billrates
xlist <- getYieldCurve("https://data.treasury.gov/feed.svc/DailyTreasuryBillRateData?$filter=month(INDEX_DATE)\%20eq\%2012\%20and\%20year(INDEX_DATE)\%20eq\%202020")
xlist
           ROUND_B1_CLOSE_4WK_2 ROUND_B1_YIELD_4WK_2 ROUND_B1_CLOSE_8WK_2 ROUND_B1_YIELD_8WK_2
2020-12-01                 0.07                 0.07                 0.07                 0.07
2020-12-02                 0.07                 0.07                 0.08                 0.08
2020-12-03                 0.08                 0.08                 0.08                 0.08
2020-12-04                 0.07                 0.07                 0.07                 0.07
           ROUND_B1_CLOSE_13WK_2 ROUND_B1_YIELD_13WK_2 ROUND_B1_CLOSE_26WK_2 ROUND_B1_YIELD_26WK_2
2020-12-01                  0.09                  0.09                  0.10                  0.10
2020-12-02                  0.09                  0.09                  0.10                  0.10
2020-12-03                  0.08                  0.08                  0.09                  0.09
2020-12-04                  0.09                  0.09                  0.10                  0.10
           ROUND_B1_CLOSE_52WK_2 ROUND_B1_YIELD_52WK_2 BOND_MKT_UNAVAIL_REASON CS_4WK_CLOSE_AVG
2020-12-01                  0.12                  0.12                      NA       0.07000000
2020-12-02                  0.11                  0.11                      NA       0.07000000
2020-12-03                  0.10                  0.10                      NA       0.07333333
2020-12-04                  0.11                  0.11                      NA       0.07250000
           CS_4WK_YIELD_AVG CS_8WK_CLOSE_AVG CS_8WK_YIELD_AVG CS_13WK_CLOSE_AVG CS_13WK_YIELD_AVG
2020-12-01       0.07000000       0.07000000       0.07000000        0.09000000        0.09000000
2020-12-02       0.07000000       0.07500000       0.07500000        0.09000000        0.09000000
2020-12-03       0.07333333       0.07666667       0.07666667        0.08666667        0.08666667
2020-12-04       0.07250000       0.07500000       0.07500000        0.08750000        0.08750000
           CS_26WK_CLOSE_AVG CS_26WK_YIELD_AVG CS_52WK_CLOSE_AVG CS_52WK_YIELD_AVG CF_WEEK
2020-12-01        0.10000000        0.10000000             0.120             0.120  202049
2020-12-02        0.10000000        0.10000000             0.115             0.115  202049
2020-12-03        0.09666667        0.09666667             0.110             0.110  202049
2020-12-04        0.09750000        0.09750000             0.110             0.110  202049

# https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=longtermrate
xlist <- getYieldCurve(base = "http://data.treasury.gov/feed.svc/DailyTreasuryLongTermRateData?$filter=month(QUOTE_DATE)\%20eq\%2012\%20and\%20year(QUOTE_DATE)\%20eq\%202020")
xlist
           BC_20year Over_10_Years Real_Rate
2020-12-01      1.46          1.47     -0.39
2020-12-02      1.50          1.51     -0.39
2020-12-03      1.46          1.47     -0.42
2020-12-04      1.53          1.54     -0.38

# https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=realyield
xlist <- getYieldCurve("https://data.treasury.gov/feed.svc/DailyTreasuryRealYieldCurveRateData?$filter=month(NEW_DATE)\%20eq\%2012\%20and\%20year(NEW_DATE)\%20eq\%202020")
xlist
            TC_5YEAR  TC_7YEAR TC_10YEAR TC_20YEAR TC_30YEAR
2020-12-01 -1.300375 -1.106995 -0.889312 -0.493451 -0.299437
2020-12-02 -1.335888 -1.129040 -0.897425 -0.488592 -0.293812
2020-12-03 -1.384003 -1.176739 -0.936774 -0.511666 -0.322141
2020-12-04 -1.408367 -1.177114 -0.923189 -0.484069 -0.278824

# https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=reallongtermrate
xlist <- getYieldCurve("https://data.treasury.gov/feed.svc/DailyTreasuryRealLongTermRateAverageData?$filter=month(QUOTE_DATE)\%20eq\%2012\%20and\%20year(QUOTE_DATE)\%20eq\%202020")
xlist
            RATE
2020-12-01 -0.39
2020-12-02 -0.39
2020-12-03 -0.42
2020-12-04 -0.38
}
}
\references{
\cite{GetYieldCurve - XML content does not seem to be XML #1
\url{https://github.com/mrbcuda/ustyc/issues/1}
}

\cite{Unknown IO error #2
\url{https://github.com/mrbcuda/ustyc/issues/2}
}
}
\author{
Andre Mikulec (partial re-write)

Matt Barry (original)
}
