% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModelBase.R
\name{ifelse.xts}
\alias{ifelse.xts}
\title{If Then Else}
\usage{
ifelse.xts(test, yes, no, ...)
}
\arguments{
\item{test}{xts object}

\item{yes}{if T, select this value}

\item{no}{if F, select this value}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
If Then Else
}
}
\examples{
\dontrun{

# ifelse.xts(If Then Else) examples

xts(matrix(c(T,F),ncol=1), zoo::as.Date(0:1))
           [,1]
1970-01-01  TRUE
1970-01-02 FALSE

ifelse.xts(xts(matrix(c(T,F),ncol=1), zoo::as.Date(0:1)), 11, 10)
           V1ie
1970-01-01   11
1970-01-02   10

xts(matrix(c(T,F,F,T),ncol=2), zoo::as.Date(0:1))
            [,1]  [,2]
1970-01-01  TRUE FALSE
1970-01-02 FALSE  TRUE

ifelse.xts(xts(matrix(c(T,F,F,T),ncol=2), zoo::as.Date(0:1)), 11, 10)
           V1ie V2ie
1970-01-01   11   10
1970-01-02   10   11

ifelse.xts(xts(matrix(c(T,F,F,T),ncol=2), zoo::as.Date(0:1)), c(11,21), c(10,20))
           V1ie V2ie
1970-01-01   11   10
1970-01-02   20   21

xts(matrix(c(11,10),ncol=1), zoo::as.Date(0:1))
           [,1]
1970-01-01   11
1970-01-02   10

# note: some functions, I must test on the coredata. E.g. is.na
ifelse.xts(xts(matrix(c(11,10),ncol=1), zoo::as.Date(0:1)) > 10 , 21, 20)
           V1ie
1970-01-01   21
1970-01-02   20
}
}
