% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{INA}
\alias{INA}
\title{Perform is.na About an xts Object}
\usage{
INA(x, ...)
}
\arguments{
\item{x}{xts object}

\item{...}{dots passes}
}
\value{
xts object of T/F values
}
\description{
\preformatted{
Without the function, INA, is.na(xts) strips the index.
Probably this function should be an xts enhancement.
}
}
\examples{
\dontrun{

# INA(Perform is.na About an xts Object) examples

xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2))

           [,1]
1970-01-01   NA
1970-01-02    0
1970-01-03   NA

INA(xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2)))

           V1ina
1970-01-01  TRUE
1970-01-02 FALSE
1970-01-03  TRUE

INA(xts::xts(matrix(c(NA_real_, 0, NA_real_,0), ncol = 2), zoo::as.Date(0:1)))
           V1ina V2ina
1970-01-01  TRUE  TRUE
1970-01-02 FALSE FALSE

explode(IBM.Open.TTR.SMA.n.2, Fun = "INA")
           IBM.Open.TTR.SMA.n.2.INA
1970-01-02                     TRUE
1970-01-05                    FALSE
1970-01-06                    FALSE
1970-01-07                    FALSE
1970-01-08                    FALSE
1970-01-09                    FALSE
1970-01-12                    FALSE
}
}
