% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{initEnv}
\alias{initEnv}
\title{sets the enviroment}
\usage{
initEnv(init = NULL)
}
\value{
envi, options, and TZ are set
}
\description{
\preformatted{

This is a space-saver that is meant to be used at the beginning of a function.

Variable ops from the calling environment sets R options.
Environment variable TZ using the calling environment variable oldtz sets.

}
}
\examples{
\dontrun{
# > options(max.print=88888L)
# initEnv()
# > getOption("max.print")
# [1] 99999
}
}
