% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{interleave}
\alias{interleave}
\title{Interleave two objects of arbitrary length}
\usage{
interleave(x, y)
}
\arguments{
\item{x}{index-able objet by `[` and the same index length as y.  Compatible type with y and combinable(c)}

\item{y}{index-able object by `[` and the same index length as x. Compatible type with x and combinable(c)}
}
\value{
new combined object
}
\description{
\preformatted{

from R CRAN package rmngb
}
}
\examples{
\dontrun{
interleave(letters[1:2], LETTERS[1:2])
}
}
