% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{quantmod___is.method.available}
\alias{quantmod___is.method.available}
\alias{is.method.available}
\title{check the .libPaths() to see a package has a training function}
\usage{
quantmod___is.method.available(method, package)

is.method.available(method, package)
}
\arguments{
\item{method}{String. No Default. Required. Name of a R package function.}

\item{package}{String. No Default. Required. Name of a R package}
}
\value{
Logical. Availablity of a function in an R package
}
\description{
Extracted because the R CRAN package quantmod does not export this.
}
