% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{is.na.xts}
\alias{is.na.xts}
\title{Perform is.na About an xts Object}
\usage{
\method{is.na}{xts}(x, ...)
}
\arguments{
\item{x}{xts object}

\item{...}{dots passes}
}
\value{
xts object of T/F values
}
\description{
\preformatted{
Without this function, is.na(xts) strips the index.
Probably this function should be an xts enhancement.
}
}
\examples{
\dontrun{

is.na.xts(Perform is.na About an xts Object) examples

xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2))

           [,1]
1970-01-01   NA
1970-01-02    0
1970-01-03   NA

is.na(xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2)))

            [,1]
1970-01-01  TRUE
1970-01-02 FALSE
1970-01-03  TRUE

NOTE
is.na(xts::xts(matrix(c(NA_real_, 0, NA_real_,0), ncol = 2), zoo::as.Date(0:1)))
            V1na  V2na
1970-01-01  TRUE  TRUE
1970-01-02 FALSE FALSE

}
}
