% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{isConnectedEM}
\alias{isConnectedEM}
\title{Test The RPostgreSQL Connection}
\usage{
isConnectedEM(conn, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection}

\item{...}{Dots passed.}
}
\value{
TRUE(connected) or FALSE(otherwise)
}
\description{
Is the connection not expired or not valid?
}
\examples{
\dontrun{
isConnectedEM(conn)
}
}
