% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{last}
\alias{last}
\title{Return Last n Elements of A Data Object}
\usage{
last(x, ...)
}
\arguments{
\item{x}{1 or 2 dimensional data object.}

\item{...}{Dots passed.}
}
\value{
last elements or zero length element
}
\description{
A generic function to return the last or last elements or rows of a vector or two-dimensional data object.
This is a wrapper over the R CRAN package xts S3 function last.
Different here is that if n elements can not be returned, then instead, zero elements are returned.
}
