% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{lastly}
\alias{lastly}
\title{Tries to Return Last n Elements of A Data Object}
\usage{
lastly(x, ...)
}
\arguments{
\item{x}{1 or 2 dimensional data object.}

\item{...}{Dots passed. Typically, tries to gather up to "n" elements to return.}
}
\value{
last elements or zero length element
}
\description{
A generic function to return the last or last elements or rows of a vector or two-dimensional data object.
This is a wrapper over the R CRAN package xts S3 function last.

It tries to return (if available) up to "n" elements.
}
