% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LD}
\alias{LD}
\title{Lead an xts Object}
\usage{
LD(x, k = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{k}{choose 1 or greater  to look into the future}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
This does not complain when: any(abs(k) > NROW(xTs)):
zoo_lag can not and will not handle. So k's are eliminated
beforehand.
}
}
\examples{
\dontrun{
# LD(lead 1) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

LD(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1ld.1 V2ld.1
1970-01-01     -2     16
1970-01-02     -4     32
1970-01-03     NA     NA

# LD(lead 1 and 2) example

LD(xts(matrix(c(1,-2,-4), ncol = 1), zoo::as.Date(0:2)), k = 1:2)
           V1ld.1 V2ld.2
1970-01-01     -2     -4
1970-01-02     -4     NA
1970-01-03     NA     NA
}
}
