% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LG}
\alias{LG}
\title{Lag an xts Object}
\usage{
LG(x, k = 1, ...)
}
\arguments{
\item{k}{choose 0 or 1 or greater to peer into the current and the past}
}
\description{
\preformatted{

This does not complain when: any(abs(k) > NROW(xTs)):
zoo_lag can not and will not handle. So k's are eliminated
beforehand.

}
}
\examples{
\dontrun{

# LG(lagXts lagging 1) example

LG(IBM.Open.TTR.SMA.n.2, k = 1)
           IBM.Open.TTR.SMA.n.2.lag.1
1970-01-02                         NA
1970-01-05                         NA
1970-01-06                   18.26250
1970-01-07                   18.35625
1970-01-08                   18.41875
1970-01-09                   18.43125
1970-01-12                   18.45625
}
}
