% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{loadAndView}
\alias{loadAndView}
\title{Load and View Files}
\usage{
loadAndView(SubPathFileExt, HowFileName = "fileName", env, ...)
}
\arguments{
\item{SubPathFileExt}{Location of the file.}

\item{HowFileName}{String. Function name to determine which part (or all of the FileName) to be the new R object name.  Expected values are "fileName", "filePostFixName", or "filePreFixName".}

\item{env}{Target environment to load data.frames. Default is the global environment (.GlobalEnv)}
}
\value{
String. Name of the data.frame loaded. data.frames are loaded into the environment.  Attempted to be started is View() upon each data.frame.
}
\description{
Creates a data.frame from each  file.  Loads those data.frames into the "env" environment.  Calls View() upon each data.frame (so the user may see the data).
}
\examples{
\dontrun{
e <- new.env()
loadAndView(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
), env = e)
}
}
