% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{loadFile}
\alias{loadFile}
\title{Load File}
\usage{
loadFile(SubPathFileExt, HowFileName = "fileName", env, ...)
}
\arguments{
\item{SubPathFileExt}{Location of the file. Required.}

\item{HowFileName}{String. Function name to determine which part (or all of the FileName) to be the new R object name.  Expected values are "fileName", "filePostFixName", or "filePreFixName".}

\item{env}{Target environment to load data.frames. Default is the global environment (.GlobalEnv)}

\item{...}{Dots passed.}
}
\value{
String. Name of the data.frame loaded. data.frames are loaded into the environment "env".
}
\description{
Creates a data.frame from each file.  Loads those data.frames into the "env" environment.  Calls View() upon each data.frame (so the user may see the data).
}
\examples{
\dontrun{
e <- new.env()
loadFile(c(
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\setup.dbf",
  "C:\\\\Program Files (x86)\\\\Stock Investor\\\\Professional\\\\Static\\\\si_isq.dbf"
), env = e)

eFST <- new.env()
DIR <- "C:\\\\DATA\\\\AAIISIPRO\\\\MONTHDATE\\\\18565"
FSTFiles <- paste0( DIR, "\\\\" , dir(DIR, pattern = glob2rx("*.FST")))
loadFile(FSTFiles, env = eFST)
str(as.list(eFST))

eDBF <- new.env()
DIR <- "C:\\\\DATA\\\\AAIISIPRO\\\\MONTHDATE\\\\18565"
DBFFiles <- paste0( DIR, "\\\\" , dir(DIR, pattern = glob2rx("*.DBF")))
loadFile(DBFFiles, env = eDBF)
str(as.list(eDBF))
}
}
