% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{modelData<-}
\alias{modelData<-}
\title{low-level set the values quantmod object slots}
\usage{
modelData(x, ...) <- value
}
\arguments{
\item{x}{quantmod object}

\item{...}{list of name-value pairs}
}
\description{
\preformatted{
Set the values quantmod object slots.
}
}
\examples{
\dontrun{
data(sample_matrix, package = "xts")
modelData(specmodel) <- list(
  training.data = zoo::index(as.xts(sample.matrix))
)
}
}
