% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{modelData}
\alias{modelData}
\title{Extract Dataset Created by specifyModel}
\usage{
modelData(x, data.window = NULL, exclude.training = FALSE)
}
\arguments{
\item{x}{A quantmod object. Required.}

\item{data.window}{Default is NULL.  A character vector of subset start and end dates to return. Alternately, this can be a list of character vectors.}

\item{exclude.training}{Logical. Default is FALSE. Remove the training period.}
}
\value{
An object of class zoo containing all transformations to data specified in specifyModel.
}
\description{
\preformatted{
This is a re-implementation of the R CRAN package quantmod
function modelData. This includes the additional flexiblity that
data.window also can be a list of character vectors.

Extract from a quantmod object the dataset
created for use in modeling.

specifyModel creates a zoo object for use in subsequent
workflow stages (buildModel, tradeModel) that
combines all model inputs, from a variety of sources,
into one model frame.

modelData returns this object.
}
}
\details{
\preformatted{
When a model is created by specifyModel,
it is attached to the returned object.
One of the slots of this S4 class is model.data.
}
}
\author{
Andre Mikulec (re-implementation)

Jeffrey A. Ryan
}
