% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{multiInterleave}
\alias{multiInterleave}
\title{Interleave Multiple Objects of Arbitrary Length}
\usage{
multiInterleave(...)
}
\arguments{
\item{...}{Dots passed. Index-able objects by `[` and all objects have the same index length as each other.  All objects must have a compatible type with each other and combinable(c) with each other.}
}
\value{
new combined object
}
\description{
\preformatted{
Inspired by R CRAN package rmngb
}
}
\examples{
\dontrun{
multiInterleave(letters[1:2], LETTERS[1:2], 10:11)
}
}
\author{
Andre Mikulec
}
