% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{multiWise}
\alias{multiWise}
\title{Setwise Interleave of Many "Two Dimensional" Index-able Objects}
\usage{
multiWise(...)
}
\arguments{
\item{...}{Dots passed. Two dimension index-able objects Two or more collections of Index-able objects.  Tested with data.frames and xts objects.}
}
\value{

}
\description{
\preformatted{
If any has less index elements than the others
then recycle the index elements.
}
}
\examples{
\dontrun{
multiWise(airquality[1:2,], iris[1:2,1:2], mtcars[1:2, 1:3])
}
}
