% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVAR.R
\name{Names<-}
\alias{Names<-}
\title{garantee that Names assigns a zero length character vector}
\usage{
Names(x) <- value
}
\arguments{
\item{x}{names}

\item{value}{result}
}
\description{
\preformatted{

if x or value is null or has length zero(0) then
then that item is character(0)

}
}
