% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBasis.R
\name{Names}
\alias{Names}
\title{better Names}
\usage{
Names(x)
}
\arguments{
\item{x}{names}
}
\description{
\preformatted{

If the argument to names is NULL or has zero(0) length
then instead for returning NULL, return character(0).

}
}
