% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{nameVect}
\alias{nameVect}
\title{Create a Named Vector from a dataframe, Table or Vector}
\usage{
nameVect(x, name)
}
\arguments{
\item{x}{The source dataframe, table, vector, or factor}

\item{name}{The column name you would like to pull out as a named vector. OR the names of the vector (if x is a vector)}
}
\value{
a named vector or factor
}
\description{
\preformatted{
see ? caroline::nv

This is R cran package caroline function "nv" with modifications
JAN 2021
https://github.com/cran/caroline/blob/af201137e4a31d675849291a1c9c07a0933b85de/R/convert.R

Modifications:

* function "names" is replaced with the safer function "Names"
  Note, keeping function "Names" and not replacing using "colnames"
  keeps generalization ??
}
}
