% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{NC}
\alias{NC}
\title{Perform na.locf About an xts Object}
\usage{
NC(x, d = "f", ...)
}
\arguments{
\item{x}{xts object}

\item{d}{direction. "f"(default) forward.  The other option is "b" backward.}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Perform na.locf About an xts Object
}
}
\examples{
\dontrun{

# NC(Perform na.locf About an xts Object) example

xts(c(NA_real_, 3, NA_real_,5, NA_real_), zoo::as.Date(0:4))
           [,1]
1970-01-01   NA
1970-01-02    3
1970-01-03   NA
1970-01-04    5
1970-01-05   NA

NC(xts(c(NA_real_, 3, NA_real_,5, NA_real_), zoo::as.Date(0:4)))
           V1nc.f
1970-01-01     NA
1970-01-02      3
1970-01-03      3
1970-01-04      5
1970-01-05      5
}
}
