% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{newColName}
\alias{newColName}
\title{generate a good column name}
\usage{
newColName(
  x = NULL,
  Fun = NULL,
  isCharFun = NULL,
  x1 = NULL,
  x2 = NULL,
  FlagsCombo = NULL,
  AltName = NULL,
  asIsAltName = NULL,
  Prefix = NULL,
  FixedSep = NULL
)
}
\arguments{
\item{x}{object with the old column names}

\item{Fun}{literally written function or the function name inside a string.
if mode(Fun) == "function", then the returned (xts and) xts root column name is "Anon".
Alternately instead of "Anon", the use may (always) explicity provide the
return (xts and) xts root column name.}

\item{isCharFun}{hint as to Fun being a function or the function name inside of a string}

\item{x1}{object providing part of the new column name}

\item{x2}{object providing (yet another) part of the the new column name}

\item{FlagsCombo}{list with named elements and their values to (eventually) become part of the new column name}

\item{AltName}{column new root name.  Default is NULL. The input value is a string. Unless a string is provided then the root name will not be replaced.}

\item{asIsAltName}{Default is NULL(F). NULL is interpreted as F.
This NULL(F) means do not attempt to simplify the AltName.
For example, if the user passed "TTR::SMA" and asIsAltName == T, then "TTR::SMA" would
directly tried to become part of the column name (and "::" would cause an error so this
would not work.).  If the user passed "TTR::SMA" and asIsAltName == NULL(F), then
"SMA" would be directly tried to directly become part of the column name.}

\item{Prefix}{place the new addition to the column name at the front instead of the end. Default is NULL. Internally the default is FALSE.}

\item{FixedSep}{replacement for "[.]|::" that was found in the function name}
}
\value{
object with the new column names
}
\description{
\preformatted{

Given column names of one or two two dimensional objects,
each with just one single column,  and a function and its arguments,
generate a new column name and return the original 'one column' two dimensional object

}
}
