% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{newXtsColName}
\alias{newXtsColName}
\title{generate a good column name}
\usage{
newXtsColName(
  xTs = NULL,
  Fun = NULL,
  isCharFun = NULL,
  xTs1 = NULL,
  xTs2 = NULL,
  WhichCombo = NULL,
  AltName = NULL,
  Prefix = NULL,
  FixedSep = NULL
)
}
\arguments{
\item{x}{object with the old column names}
}
\value{
object with the old column names
}
\description{
\preformatted{

}
}
