% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{NVAR}
\alias{NVAR}
\alias{NVAR.default}
\alias{NVAR.xts}
\title{number of variables}
\usage{
NVAR(x = NULL)

\method{NVAR}{default}(x = NULL)

\method{NVAR}{xts}(x = NULL)
}
\arguments{
\item{x}{object}
}
\value{
integer number of variables
}
\description{
\preformatted{

NCOL wrongly returns value one(1)
on non-data.frame 2nd dimension objects
with a 2nd dimension size of zero(0).

}
}
\examples{
\dontrun{
# > require(xts)
# > NVAR(xts(, zoo::as.Date("1970-01-12")))
# [1] 0
}
}
