% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{objectNameFixEM}
\alias{objectNameFixEM}
\title{Format input for database schema/object names.}
\usage{
objectNameFixEM(
  connName,
  o.nm,
  as.identifier = TRUE,
  dbQuote = "Identifier",
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".
where a dummy connection will be used.}

\item{o.nm}{Object name string, length 1-2.}

\item{as.identifier}{Boolean whether to return (schema,table) name as database
sanitized identifiers (TRUE) or as regular character (FALSE)}

\item{dbQuote}{String. Only used when "as.identifier = TRUE". Default is "Identifier". Alternately, this value can be "Literal."}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
character vector of length 2. Each character element is in
    (escaped) double-quotes when as.identifier = TRUE.
}
\description{
This is a near copy of the R CRAN package rpostgis function dbTableNameFix.
}
\details{
Internal rpostgis function to return common (length = 2) schema
and table name vector from various table and schema + table name
inputs.
}
\examples{
\dontrun{
name <- c("schema","table")
objectNameFixEM(o.nm = name)

#current search path schema is added to single-length character object (if only table is given)
name<-"table"
objectNameFixEM(o.nm = name)

#schema or table names with double quotes should be given exactly as they are
(make sure to wrap in single quotes in R):
name <- c('sch"ema','"table"')
objectNameFixEM(o.nm = name)
}
}
\keyword{internal}
