% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\alias{.onUnload}
\title{.onLoad getOption package settings}
\usage{
.onLoad(libname, pkgname)

.onUnload(libpath)
}
\arguments{
\item{libname}{defunct}

\item{pkgname}{defunct}
}
\value{
invisible()

invisible()
}
\description{
.onLoad getOption package settings

.onLoad getOption package settings
}
\examples{
\dontrun{
# See all options
options()

# See the current temporary working folder
getOption("econModel.tryCatchLog.write.error.dump.folder"))

# Of my manually self-pre-created folder . . .
# Re-assign to my permanent(ish) working folder
MyWorkingFolder <- "C:\\\\Users\\\\Public\\\\econModelWork"
MyWorkingFolder <- normalizePath(MyWorkingFolder, winslash = "/")
options(econModel.tryCatchLog.write.error.dump.folder = MyWorkingFolder)
getOption("econModel.tryCatchLog.write.error.dump.folder"))
}
getOption("econModel.name")
}
