% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{pairWise}
\alias{pairWise}
\title{pairwise interleave of two two dimensional index-able objects}
\usage{
pairWise(x, y)
}
\arguments{
\item{x}{two dimension index-able object}

\item{y}{two dimension index-able object}
}
\value{
list of length two of two interleaved two dimensional index-able  objects
}
\description{
\preformatted{

If one or the other has one index element while the other
has N index elements, then the one will be recycled to N index elements.

}
}
\examples{
\dontrun{
#
 list(iris[1:2,1:2], airquality[1:2,1:2])
# [[1]]
#   Sepal.Length Sepal.Width
# 1          5.1         3.5
# 2          4.9         3.0
#
# [[2]]
#   Ozone Solar.R
# 1    41     190
# 2    36     118
#
 str( pairWise( iris[1:2,1:2], airquality[1:2,1:2] ) )
# List of 2
#  $ :List of 2
#   ..$ Sepal.Length: num [1:2] 5.1 4.9
#   ..$ Ozone       : int [1:2] 41 36
#  $ :List of 2
#   ..$ Sepal.Width: num [1:2] 3.5 3
#   ..$ Solar.R    : int [1:2] 190 118
#
# # 1 by N recycling
# #
  str( pairWise( iris[1:2,1:2], airquality[1:2,1, drop = F] ) )
#  List of 2
#   $ :List of 2
#    ..$ Sepal.Length: num [1:2] 5.1 4.9
#    ..$ Ozone       : int [1:2] 41 36
#   $ :List of 2
#    ..$ Sepal.Width: num [1:2] 3.5 3
#    ..$ Ozone      : int [1:2] 41 36
#
 library(xts)
 data("sample_matrix", package = "xts")
 str( pairWise(as.xts(sample_matrix)[,1:2], as.xts(sample_matrix)[,3:4] ) )
# List of 2
#  $ :List of 2
#   ..$ Open:An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50 50.2 50.4 50.4 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Open"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#   ..$ Low :An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50 50.2 50.3 50.2 50.1 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Low"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#  $ :List of 2
#   ..$ High :An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50.1 50.4 50.4 50.4 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "High"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#   ..$ Close:An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50.1 50.4 50.3 50.3 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Close"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#
 library(quantmod)
 ibm <- quantmod::getSymbols("IBM", from = "1970-01-01", to = "1970-01-13", auto.assign = FALSE)
 pairWise(tail(ibm[,c("IBM.Open","IBM.Close")]), xts(, zoo::as.Date(0)[0]))

# [[1]]
# [[1]]$IBM.Open
# IBM.Open
# 1970-01-05   18.300
# 1970-01-06   18.413
# 1970-01-07   18.425
# 1970-01-08   18.438
# 1970-01-09   18.475
# 1970-01-12   18.450
#
# [[1]][[2]]
# Data:
#   numeric(0)
#
# Index:
#   integer(0)
#
# [[2]]
# [[2]]$IBM.Close
# IBM.Close
# 1970-01-05    18.413
# 1970-01-06    18.425
# 1970-01-07    18.438
# 1970-01-08    18.475
# 1970-01-09    18.450
# 1970-01-12    18.388
#
# [[2]][[2]]
# Data:
#   numeric(0)
#
# Index:
#   integer(0)
#
 pairWise(tail(ibm[,c("IBM.Open")]), xts())
# [[1]]
# [[1]]$IBM.Open
# IBM.Open
# 1970-01-05   18.300
# 1970-01-06   18.413
# 1970-01-07   18.425
# 1970-01-08   18.438
# 1970-01-09   18.475
# 1970-01-12   18.450
#
# [[1]][[2]]
# Data:
#   numeric(0)
#
# Index:
#   integer(0)
#
#
}
}
