% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{partBoundEM}
\alias{partBoundEM}
\title{From a PostgreSQL Object Get Its Partition Bounds}
\usage{
partBoundEM(conn, DfName)
}
\arguments{
\item{conn}{PostgreSQLConnection.}

\item{DfName}{String. Name of the PostgreSQL object.}
}
\value{
vector of partiton bound list values.
}
\description{
Of a PostgreSQL inheriting partition, detect its partition bounds.
If not a "inheriting partition", an empty character vector is returned.
}
\details{
This is only designed to work on List paritions the non-expression partition bounds.
}
