% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{predictModel.default}
\alias{predictModel.default}
\title{Predict}
\usage{
predictModel.default(object, data, ...)
}
\arguments{
\item{object}{R Object. Default is none. Required. Machine Learning object.}

\item{data}{New Data.  Default is none. Required.}

\item{...}{Dots Passed.}
}
\value{
Prediction.
}
\description{
Given an R Object and new data make predictions
}
