% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{predictModel.train}
\alias{predictModel.train}
\title{Predict}
\usage{
predictModel.train(object, data, ...)
}
\arguments{
\item{object}{R Object. Default is none. Required. Machine Learning object.}

\item{data}{New Data.  Default is none. Required.}

\item{...}{Dots Passed.}
}
\value{
Prediction.
}
\description{
Predict using the R CRAN package Caret class "train"
}
