% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModelBase.R
\name{prntPrpr}
\alias{prntPrpr}
\title{Print Proper (and short)}
\usage{
prntPrpr(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots passed}
}
\value{
x object or "print proper" string
}
\description{
\preformatted{
Print Proper (and short)
}
}
\examples{
\dontrun{

# prntPrpr(Print Proper (and short)) example

prntPrpr(T)
[1] "T"

prntPrpr(c(T,F,NA))
[1] "T" "F" "N"
}
}
