% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantmod___convert.time.series.R
\name{quantmod___convert.time.series}
\alias{quantmod___convert.time.series}
\title{Convert a time series to an R CRAN package quantmod supported time series}
\usage{
quantmod___convert.time.series(fr, return.class)
}
\arguments{
\item{fr}{see R CRAN package quantmod private function convert.time.series}

\item{return.class}{see R CRAN package quantmod private function convert.time.series}
}
\value{
R CRAN quantmod compatible time series
}
\description{
Used internally by the function getSymbols.ALFRED
}
\details{
Replica of quantmod:::convert.time.series

Code is repeated here because it is private code in the R CRAN package quantmod.  CRAN rules do not allow private functions in other packages to be called directly.
}
