% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModelSupport.R
\name{R.utils__copyDirectoryByPattern}
\alias{R.utils__copyDirectoryByPattern}
\title{copy Directory using a Regular Expression}
\usage{
R.utils__copyDirectoryByPattern(
  from,
  to = ".",
  ...,
  private = TRUE,
  recursive = TRUE,
  pattern = NULL,
  tolower = FALSE
)
}
\arguments{
\item{from}{The pathname of the source directory to be copied.}

\item{to}{The pathname of the destination directory.}

\item{...}{Additional arguments passed to
    \code{\link[base:files]{file.copy}}(), e.g. \code{overwrite}.}

\item{private}{If \code{\link[base:logical]{TRUE}}, files (and directories) starting with
        a period is also copied, otherwise not.}

\item{recursive}{If \code{\link[base:logical]{TRUE}}, subdirectories are copied too,
        otherwise not.}

\item{tolower}{default is F. Convert the target file names to lower case.}
}
\value{
Returns (invisibly) a \code{\link[base]{character}} \code{\link[base]{vector}} of pathnames copied.
}
\description{
\preformatted{
Code copy of R.utils::copyDirectory.default with the addition of the
parameter (and feature) "tolower".  The function is renamed to
be R.utils__copyDirectoryByPattern
}
}
\details{
Note that this method does \emph{not} use \code{\link[R.utils]{copyFile}}() to
  copy the files, but \code{\link[base]{file.copy}}().
}
\examples{
\dontrun{
}
}
