% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModelSupport.R
\name{regExtract}
\alias{regExtract}
\title{Find All Substrings}
\usage{
regExtract(pattrn, x)
}
\arguments{
\item{pattrn}{String. Regular Expression Pattern.}

\item{x}{vector of Strings.}
}
\value{
vectors of found substrings
}
\description{
Given a string or vector of strings, find and extract all substrings that match the pattern.
}
\examples{
\dontrun{
# not including the parentheses
regExtract("(?<=\\\\()(.*?)(?=\\\\))", "LIST (c1)")
[1] "c1"

regExtract("(?<=\\\\()(.*?)(?=\\\\))", "LIST (((c1)::integer))")
[1] "((c1"

RegExtract("(?<=\\\\()(\\\\w+)(?=\\\\))", "LIST (((c1)::integer))")
[1] "c1"
}
}
\references{
\cite{Regular expression to extract text between square brackets
\url{https://stackoverflow.com/questions/2403122/regular-expression-to-extract-text-between-square-brackets}
}
}
