% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{relativeRows}
\alias{relativeRows}
\title{Return the Correct Elements of A Data Object}
\usage{
relativeRows(x, ...)
}
\arguments{
\item{x}{1 or 2 dimensional data object.}

\item{...}{Dots passed.}
}
\value{
first elements or zero length element
}
\description{
A function to return the rows of a vector or two-dimensional data object.
This is a wrapper over the R CRAN package econModel first and last.
Different here is that if n elements can not be returned, then instead, stop() happens.
}
