% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{relChg}
\alias{relChg}
\title{Relative Change}
\usage{
relChg(x = NULL, base = 0, lag = 1, log = FALSE, ...)
}
\arguments{
\item{x}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}

\item{log}{logrithmic}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Relative Change
}
}
\examples{
\dontrun{

# relChg (relative change) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

relChg(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1rc.0.1 V2rc.0.1
1970-01-01       NA       NA
1970-01-02       -2        2
1970-01-03       -1        2
}
}
