% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RN}
\alias{RN}
\title{Time Near Events or Event Run Identification}
\usage{
RN(x, z = "a", e = T, n = F, ...)
}
\arguments{
\item{x}{xts object}

\item{z}{"after"(default). Before or after the event. "before" is the other option.}

\item{e}{T(default). Detect the observation that holds the value of T(TRUE). This can be any value that can be testing on the right hand side of the equals (==) sign. A special option to be an event is NA.}

\item{n}{F(default). Do not "!" not-the-event.  T is useful against NA to detect !is.na(NA). The special case of NA internally is handled !is.na(NA).}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Counts the number of observations to/from an event.
This is called the time-distance to/from an event.
Alternately, gives the range of an event.

Range of the event - Run - (RN)
}
}
\examples{
\dontrun{

# RN(Time Near Events or Event Run Identification) examples
}
}
