% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RPCP}
\alias{RPCP}
\title{Relative Proportional Change in Percent}
\usage{
RPCP(x, b = 0, l = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{b}{choose -1 (or less) to look into the future}

\item{l}{observations backwards}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Ninety-nine (99%) percent of the people in the world
are using THIS WRONG ONE.
Relative Proportional Change(RPC) multiplied by 100.00
Relative Proportional Change(RPC) in percent (RPCP)
}
}
\examples{
\dontrun{

RPCP(Relative Proportional Change in Percent) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

RPCP(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1rpcp.0.1 V2rpcp.0.1
1970-01-01         NA         NA
1970-01-02       -300        100
1970-01-03        100        100
}
}
