% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RRC}
\alias{RRC}
\title{Relative Relative Change}
\usage{
RRC(x = NULL, b = 0, l = 1, lg = FALSE, ...)
}
\arguments{
\item{x}{xts object}

\item{b}{choose -1 (or less) to look into the future}

\item{l}{observations backwards}

\item{lg}{log}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Absolute value of "Absolute Relative Change"
Results NEVER include "negative numbers."
}
}
\examples{
\dontrun{

# RRC (Relative Relative Change) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

RRC(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1rrc.0.1 V2rrc.0.1
1970-01-01        NA        NA
1970-01-02         2         2
1970-01-03         1         2
}
}
