% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPeriod.R
\name{seq.yearmon}
\alias{seq.yearmon}
\title{zoo seq.yearmon}
\usage{
\method{seq}{yearmon}(from, to, by, length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{See ? seq.Date}

\item{to}{See ? seq.Date}

\item{by}{See ? seq.Date. See examples (below).}

\item{length.out}{See ? seq.Date}

\item{along.with}{See ? seq.Date}

\item{...}{See ? seq.Date}
}
\value{
zoo as.yearmon class
}
\description{
\preformatted{
Originally meant to be a zoo S3 object.
Based on the code of seq.Date.

The situation of "no existing seq.yearmon" drove me nuts!
Therefore I programmed one.

Currently, unlike seq.Date, this function
does not have a difftime implementation.

Currently, unlike seq.Date, this function
does not try to convert to POSIX__.

NOTE: this implementations compliments the gist by adding "by methods"
See. the examples.

}
}
\examples{
\dontrun{

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"))
# Error in seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"))  :
#  exactly two of 'to', 'by' and 'length.out' / 'along.with' must be specified
# That is the expected output!

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), length.out = 2)
# [1] "Jan 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), length.out = 3)
# [1] "Jan 2000" "Apr 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), length.out = 4)
# [1] "Jan 2000" "Mar 2000" "May 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), by = 1/12)
# [1] "Jan 2000" "Feb 2000" "Mar 2000" "Apr 2000" "May 2000" "Jun 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), by = 2/12)
# [1] "Jan 2000" "Mar 2000" "May 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), by = "2 yearmons")
# [1] "Jan 2000" "Mar 2000" "May 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-01"), to = as.yearmon("2000-07"), by = "2 months")
# [1] "Jan 2000" "Mar 2000" "May 2000" "Jul 2000"

seq.yearmon(as.yearmon("2000-07"), to = as.yearmon("2000-01"), by = -1/12)
# [1] "Jul 2000" "Jun 2000" "May 2000" "Apr 2000" "Mar 2000" "Feb 2000" "Jan 2000"

seq.yearmon(as.yearmon("2000-07"), to = as.yearmon("2000-01"), by = "-1 yearmons")
# [1] "Jul 2000" "Jun 2000" "May 2000" "Apr 2000" "Mar 2000" "Feb 2000" "Jan 2000"

}
}
\references{
\cite{R zoo S3 object seq.yearqtr seq.yearmon
\url{https://gist.github.com/AndreMikulec/aceb20a0b6c170027b035519ca7a3adb}}
}
