% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPeriod.R
\name{seq.yearqtr}
\alias{seq.yearqtr}
\title{zoo seq.yearqtr}
\usage{
\method{seq}{yearqtr}(from, to, by, length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{See ? seq.Date}

\item{to}{See ? seq.Date}

\item{by}{See ? seq.Date. See examples (below).}

\item{length.out}{See ? seq.Date}

\item{along.with}{See ? seq.Date}

\item{...}{See ? seq.Date}
}
\value{
zoo as.yearqtr class
}
\description{
\preformatted{
Originally meant to be a zoo S3 object.
Based on the code of seq.Date.

The situation of "no existing seq.yearqtr" drove me nuts!
Therefore I programmed one.
Currently, unlike seq.Date, this function
does not have a difftime implementation.

Currently, unlike seq.Date, this function
does not try to convert to POSIX__.

NOTE: this implementations compliments the gist by adding "by methods"
See. the examples.

}
}
\examples{
\dontrun{
seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"))
# Error in seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1")) :
#  exactly two of 'to', 'by' and 'length.out' / 'along.with' must be specified
# That was the expected output!

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), length.out = 2)
# [1] "2000 Q1" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), length.out = 3)
# [1] "2000 Q1" "2001 Q1" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), length.out = 4)
# [1] "2000 Q1" "2000 Q4" "2001 Q2" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), by = 1/4)
# [1] "2000 Q1" "2000 Q2" "2000 Q3" "2000 Q4" "2001 Q1" "2001 Q2" "2001 Q3"
# [8] "2001 Q4" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), by = 2/4)
# [1] "2000 Q1" "2000 Q3" "2001 Q1" "2001 Q3" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), by = "2 yearqtrs")
# [1] "2000 Q1" "2000 Q3" "2001 Q1" "2001 Q3" "2002 Q1"

seq.yearqtr(as.yearqtr("2000 Q1"), to = as.yearqtr("2002 Q1"), by = "2 quarters")
# [1] "2000 Q1" "2000 Q3" "2001 Q1" "2001 Q3" "2002 Q1"

seq.yearqtr(as.yearqtr("2002 Q1"), to = as.yearqtr("2000 Q1"), by = -1/4)
# [1] "2002 Q1" "2001 Q4" "2001 Q3" "2001 Q2" "2001 Q1" "2000 Q4" "2000 Q3"
# [8] "2000 Q2" "2000 Q1"

seq.yearqtr(as.yearqtr("2002 Q1"), to = as.yearqtr("2000 Q1"), by = "-1 yearqtrs")
# [1] "2002 Q1" "2001 Q4" "2001 Q3" "2001 Q2" "2001 Q1" "2000 Q4" "2000 Q3"
# [8] "2000 Q2" "2000 Q1"

}
}
\references{
\cite{R zoo S3 object seq.yearqtr seq.yearmon
\url{https://gist.github.com/AndreMikulec/aceb20a0b6c170027b035519ca7a3adb}}
}
