% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StressIndex.R
\name{stLouisFEDdaily2Weekly}
\alias{stLouisFEDdaily2Weekly}
\title{Convert an xts object Into a St Louis FED Weekly xts Object.}
\usage{
stLouisFEDdaily2Weekly(x, Level)
}
\arguments{
\item{x}{xts object of Dailies.}

\item{Level}{String. Default is "Levels" Other choices are "ChangeInLevels" and "ChangeInLogLevels."}
}
\value{
Modified xts object
}
\description{
Given Dailies, return the five(5) day moving average (called a "weekly") with an Observation data on Fridays that are published on the following Thursdays.
}
\examples{
\dontrun{
DGS2 <- quantmod::getSymbols("DGS2", src = "FRED", auto.assign = F)
DGS2 <- stLouisFEDdaily2Weekly(DGS2, Level = "ChangeInLevels")
}
}
