% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{subsetIndex}
\alias{subsetIndex}
\title{Create a New Zoo/Xts Index of a Lesser Count of Values}
\usage{
subsetIndex(x, Restriction)
}
\arguments{
\item{x}{zoo or xts object. Required.}

\item{Restriction}{R list of time range vectors.  Each vector has two elements of either zoo or xts index elements or index numbers. The left element is the earlier time.  The right vector is the later time.}
}
\value{
index elements
}
\description{
In Restriction enter a vector of begin/end pairs (or a list of such pairs).
}
