% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TBE}
\alias{TBE}
\title{Time Near Events or Event Run Identification}
\usage{
TBE(x, e = T, n = F, r = F, ...)
}
\arguments{
\item{x}{xts object}

\item{e}{T(default). Detect the observation that holds the value of T(TRUE). This can be any value that can be testing on the right hand side of the equals (==) sign. A special option to be an event is NA.}

\item{n}{F(default). Do not "!" not-the-event.  T is useful against NA to detect !is.na(NA). The special case of NA internally is handled !is.na(NA).}

\item{r}{F(default). Do not generate run range identifiers, instead return(T) the preferred time-distance from the event.}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Counts the number of observations to/from an event.
This is called the time-distance to/from an event.
Alternately, gives the range of an event.

Time Before Event (TBE)
}
}
\examples{
\dontrun{

# TBE(Time Near Events or Event Run Identification) examples
}
}
