% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{timeEvent}
\alias{timeEvent}
\title{Time Near Events or Event Run Identification}
\usage{
timeEvent(x, zone = "after", event = T, not = F, run = F, ...)
}
\arguments{
\item{x}{xts object}

\item{zone}{"after"(default). Before or after the event. "before" is the other option.}

\item{event}{T(default). Detect the observation that holds the value of T(TRUE). This can be any value that can be testing on the right hand side of the equals (==) sign. A special option to be an event is NA.}

\item{not}{F(default). Do not "!" not-the-event.  T is useful against NA to detect !is.na(NA). The special case of NA internally is handled !is.na(NA).}

\item{run}{F(default). Do not generate run range identifiers, instead return(T) the preferred time-distance from the event.}

\item{...}{dots passed}
}
\value{
xts object
}
\description{
\preformatted{
Counts the number of observations to/from an event.  This is called the time-distance to/from an event.
Alternately, gives the range of an event.
}
}
\examples{
\dontrun{

# timeEvent(Time Near Events or Event Run Identification) examples

xts(c(NA,F,F,T,NA,T,F,F,NA), zoo::as.Date(0:8))
            [,1]
1970-01-01    NA
1970-01-02 FALSE
1970-01-03 FALSE
1970-01-04  TRUE
1970-01-05    NA
1970-01-06  TRUE
1970-01-07 FALSE
1970-01-08 FALSE
1970-01-09    NA

timeEvent(xts(c(NA,F,F,T,NA,T,F,F,NA), zoo::as.Date(0:8)))

           V1te
1970-01-01   NA
1970-01-02   NA
1970-01-03   NA
1970-01-04    0
1970-01-05    1
1970-01-06    0
1970-01-07    1
1970-01-08    2
1970-01-09    3

timeEvent(xts(c(NA,F,F,T,NA,T,F,F,NA), zoo::as.Date(0:8)), run = T)

           V1te
1970-01-01   NA
1970-01-02   NA
1970-01-03   NA
1970-01-04    1
1970-01-05    1
1970-01-06    2
1970-01-07    2
1970-01-08    2
1970-01-09    2

timeEvent(xts(c(NA,F,F,T,NA,T,F,F,NA), zoo::as.Date(0:8)), zone = "before")

           V1te
1970-01-01    3
1970-01-02    2
1970-01-03    1
1970-01-04    0
1970-01-05    1
1970-01-06    0
1970-01-07   NA
1970-01-08   NA
1970-01-09   NA

timeEvent(xts(c(NA,F,F,T,NA,T,F,F,NA), zoo::as.Date(0:8)), zone = "before", run = T)

           V1te
1970-01-01    2
1970-01-02    2
1970-01-03    2
1970-01-04    2
1970-01-05    1
1970-01-06    1
1970-01-07   NA
1970-01-08   NA
1970-01-09   NA
}
}
