% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPeriod.R
\name{toPeriod}
\alias{toPeriod}
\title{Convert a Date time series to a series of another frequency}
\usage{
toPeriod(
  x,
  Period = "months",
  k = 1L,
  PeriodStart = NULL,
  FunEach = identity,
  FunAll = NC,
  KeepOrigValues = F,
  fillInterior = F,
  fillInteriorBy = NULL,
  Calendar = "UnitedStates/GovernmentBond",
  BusDayConv = 0L
)
}
\arguments{
\item{x}{xts object}

\item{Period}{Period to convert to. Default is "months".
See ? seq.POSIXt: "secs", "mins", "hours", "days", "weeks", "months", "quarters" or "years"
This is the aggregation (summary).}

\item{k}{Integer. Default is 1L.  Number of k Periods.}

\item{PeriodStart}{Date time. Default is NULL.  Required. Date time of the beginning of the series. Must be convertible by S3 to a POSIXct by as.POSIXct(PeriodEnd). The user must be sure that his value is less than or equal to the earliest index value of x.}

\item{FunEach}{Function. Has one argument: x; that is the xts object of the period. Default is identity. Function to be applied per period.  The value at the, per period, "last" positions is the "end of period"(EOP).}

\item{FunAll}{Function. Default is econModel::NC. Has two argements: x; that is the xts object of all periods; EOPIndex is the index of EOP index values. This function is to be applied across all periods.}

\item{KeepOrigValues}{Logical. Default is FALSE. Only the New Generated elements are kept and the original xts(x) elements are discarded from the output.  Otherwise, the original xts(x) elements are kept (and the New Generated elements are kept).}

\item{fillInterior}{Logical. Default is FALSE. Created sub(or super)-Period data points.}

\item{fillInteriorBy}{String. Default is NULL.  If fillInterior is TRUE, then this value must be provided. See seq.POSIXt: "secs", "mins", "hours", "days", "weeks", "months", "quarters" or "years".  Note, this is applied JUST BEFORE the function FunAll is called.}

\item{Calendar}{Default is "UnitedStates/GovernmentBond". Calendar to use.  See the Details section of ?? RQuantLib::Calendars.}

\item{BusDayConv}{Integer.  Default is 0L.  See \url{https://www.quantlib.org/reference/group__datetime.html} See ? RQuantLib::Enum, ? RQuantLib::adjust, and the parameter bcd(Business Day Convention). 0L means if the Day falls on a Holiday (Holiday includes weekends), then Following: the first business day after the given holiday becomes the (new) adjusted date.}
}
\value{
modified xts object
}
\description{
This is a smart wrapper over seq.POSIXt.
}
\examples{
\dontrun{
x <- xts::xts(c(3,363), zoo::as.Date(c(3,363)))
toPeriod(x, Period = "weeks", PeriodStart = zoo::as.Date(2))
toPeriod(x, Period = "weeks", k = 2L, PeriodStart = zoo::as.Date(2))
toPeriod(x, Period = "weeks", k = 2L, PeriodStart = zoo::as.Date(2), fillInterior = T, fillInteriorBy = "days")
toPeriod(x, Period = "months", k = 2L, PeriodStart = zoo::as.Date(2))
toPeriod(x, Period = "quarters", k = 2L, PeriodStart = zoo::as.Date(-20))
}
}
