% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{uninitEnv}
\alias{uninitEnv}
\title{unsets the enviroment}
\usage{
uninitEnv()
}
\value{
options, and TZ are un-set
}
\description{
\preformatted{

This is a space-saver that is meant to be used at the beginning of a function.

Variable ops from the calling environment resets R options.
Environment variable TZ using the calling environment variable oldtz resets.

}
}
\examples{
\dontrun{
# > uninitEnv()
# getOption("digits")
# [1] 5
}
}
