% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/econModelSupport.R
\name{unlockEnvironment}
\alias{unlockEnvironment}
\alias{lsNamespaceInfo}
\alias{AllInfoNS}
\alias{forceAssignInNamespace}
\title{Unlock an R Package Enviroment and Assign an Object to the R Package Environment}
\usage{
unlockEnvironment(env)

lsNamespaceInfo(ns, ...)

AllInfoNS(ns)

forceAssignInNamespace(x, value, namespace)
}
\arguments{
\item{env}{Environment. Default none. Required.}

\item{ns}{String. Default none. Required. Name of a namspace}

\item{...}{dots.  Passed to ls}

\item{x}{String. Required. Default none. Name of the symbol/function}

\item{value}{R Object. Default none. Required. New value of the parameter x symbol/function}

\item{namespace}{String. Default none. Required. Name of the namespace to assign parameter x into.}
}
\value{
Of unlockEnvironment, Logical.  TRUE if this package environment has been successfully unlocked. FALSE otherwise.

Of lsNamespaceInfo, return a character vector of objects in a namespace

Of AllInfoNS, return lists and sublists of information about the namespace

Of forceAssignInNamespace, silently assign object x (from parameter x) intoto the namespace non-exported objects collection
}
\description{
Of unlockEnvironment, unlock a non-base package environment.  Note: R package base is unlocked (by design)

Of lsNamespaceInfo, from the name of a namespace, return a character vector of all objects

Of AllInfoNS, from the name of a namespace, return information

Of forceAssignInNamespace, force the assignment of a new item into a namespace.
Adapted from Rcpp. Based on the original function forceAssignMyNamespace by Willem Ligtenberg.
}
\examples{
\dontrun{
#
# something not-on-the-search() path
#
requireNamespace("RPostgreSQL")

getNamespace("RPostgreSQL")
<environment: namespace:RPostgreSQL>

environmentIsLocked(getNamespace("RPostgreSQL"))
[1] TRUE

unlockEnvironment(getNamespace("RPostgreSQL"))
[1] TRUE

forceAssignInNamespace("prnt", function(x) {print(x)}, namespace = "RPostgreSQL")
RPostgreSQL::prnt("HelloEveryWhere")
Error: 'prnt' is not an exported object from 'namespace:RPostgreSQL'

RPostgreSQL:::prnt("HelloEveryWhere")
[1] "HelloEveryWhere"

assign("prnt","prnt", envir = AllInfoNS("RPostgreSQL")$exports)
RPostgreSQL::prnt("HelloEveryWhere")
[1] "HelloEveryWhere"
}
\dontrun{
#
# something on the search() path
#
library(RSQLite)

getNamespace("RSQLite")
<environment: namespace:RSQLite>

environmentIsLocked(getNamespace("RSQLite"))
[1] TRUE

unlockEnvironment(asNamespace("RSQLite"))
[1] TRUE

forceAssignInNamespace("prnt", function(x) {
  print(x)
}, namespace = "RSQLite")

RSQLite::prnt("HelloEveryWhere")
Error: 'prnt' is not an exported object from 'namespace:RSQLite'

RSQLite:::prnt("HelloEveryWhere")
[1] "HelloEveryWhere"

# promote that function to be exported
assign("prnt","prnt", envir = AllInfoNS("RSQLite")$exports)
RSQLite::prnt
function(x) {print(x)}

RSQLite::prnt("HelloEveryWhere")
[1] "HelloEveryWhere"
}
}
\references{
\cite{unlockEnvironment function
\url{https://github.com/openanalytics/Rango/blob/adc99e077b71c8c6826cabb7ff1266050898718a/Rango/src/unlockEnvironment.cpp}
}

\cite{envir.c in the R 2.15.1 source (still in R 3.3.1)
\url{https://github.com/SurajGupta/r-source/blob/a28e609e72ed7c47f6ddfbb86c85279a0750f0b7/src/main/envir.c}
}

\cite{forceAssignMyNamespace by Willem Ligtenberg
\url{https://github.com/openanalytics/Rango/blob/adc99e077b71c8c6826cabb7ff1266050898718a/Rango/R/utils.R}
}
}
\author{
Willem Ligtenberg is the author of unlockEnvironment

Andre Mikulec

Andre Mikulec is the author of AllInfoNS

Willem Ligtenberg is the author of forceAssignInNamespace/forceAssignMyNamespace
}
