% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{unWeave}
\alias{unWeave}
\title{Setwise unWeave of Many "Two Dimensional" Index-able Objects}
\usage{
unWeave(...)
}
\arguments{
\item{...}{Dots passed. Tested with Arguments of data.frames and xts objects.}
}
\value{
List. Same length as the number of function call input arguments.
}
\description{
\preformatted{
The number of Argument (sub)elements must be the same
}
}
\examples{
\dontrun{
colnames(iris)[1:4]
[1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"
unWeave(iris[1:2, c(1,3)], iris[1:2, c(2,4)])
}
}
