% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{viewObject}
\alias{viewObject}
\title{View Object}
\usage{
viewObject(env, ...)
}
\arguments{
\item{env}{Target environment or (list of environments) to view R objects. Required.}

\item{...}{Dots passed.}
}
\value{
None.
}
\description{
From the "env" environment.  Calls View() upon each R object.
}
\examples{
\dontrun{
e <- new.env()
assign("iris", iris, envir = e)
assign("airquality", airquality, envir = e)
viewObject(e)
}
}
