% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AC}
\alias{AC}
\title{Absolute Change}
\usage{
AC(x, b = 0, l = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{b}{choose -1 (or less) to look into the future}

\item{l}{observations backwards}

\item{...}{dots passed to lagXts}
}
\value{
xts object
}
\description{
\preformatted{
Absolute Change
}
}
\examples{
\dontrun{

# AC(absolute change) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32
#'
AC(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1ac.0.1 V2ac.0.1
1970-01-01       NA       NA
1970-01-02       -3        8
1970-01-03       -2       16
}
}
