% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{copyAAIISIProDBFs}
\alias{copyAAIISIProDBFs}
\title{Copy SIPro .DBF Files to a Target Directory}
\usage{
copyAAIISIProDBFs(
  From = "C:/Program Files (x86)/Stock Investor/Professional",
  To = tempdir()
)
}
\arguments{
\item{From}{string. Directory of AAII StockInvestor Pro installation directory.  Defaults to "C:/Program Files (x86)/Stock Investor/Professional".}

\item{To}{string. Location of the target directory. Default is in the TEMPDIR.}
}
\value{
string. Date of the SIPro update, in days since the UNIX epoch (birthday of UNIX: January 1st, 1970). Returned is the "Current as of date" of StockInvestor Pro.  This is the same data found by doing Help -> About (and then reading the bottom line).
}
\description{
AAII StockInvestor Pro source corporate and price data is stored in .DBF files.
Along with the copied .DBF files also the ReadMe.txt and the si.chm help files is copied.
}
\examples{
\dontrun{

# example of copyAAIISIProDBFs(Copy SIPro .DBF Files to a Target Directory)

copyAAIISIProDBFs(
    from = "C:/Program Files (x86)/Stock Investor/Professional"
  , to   = paste0("W:/AAIISIProDBFs","/",getAAIISIProDate())
)
}
}
