% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSorting.R
\name{cSort}
\alias{cSort}
\alias{cSort.default}
\alias{cSort.character}
\alias{cSort.numeric}
\alias{cSort.Date}
\alias{cSort.yearmon}
\alias{cSort.yearqtr}
\alias{cSort.POSIXct}
\alias{cSort.chron}
\title{custom sorting a vector}
\usage{
cSort(x, ...)

\method{cSort}{default}(x, ...)

\method{cSort}{character}(x, InitOrder, CI = FALSE, sortVectExc = TRUE, chopVectExc = FALSE, ...)

\method{cSort}{numeric}(x, InitOrder, CI = FALSE, sortVectExc = TRUE, chopVectExc = FALSE, ...)

\method{cSort}{Date}(x, InitOrder, ...)

\method{cSort}{yearmon}(x, InitOrder, ...)

\method{cSort}{yearqtr}(x, InitOrder, ...)

\method{cSort}{POSIXct}(x, InitOrder, ...)

\method{cSort}{chron}(x, InitOrder, ...)
}
\arguments{
\item{x}{vector  to be sorted}

\item{...}{dots passed}

\item{InitOrder}{starting custom sorting ( without the excess )}

\item{CI}{FALSE(default) whether or not Vector excess items that are
not found in InitOrder  are sorted 'not case insensitive'(TRUE) or
'not case sensitive'(FALSE).  Sorting is done by "lower()"}

\item{sortVectExc}{sort vector excess. TRUE(default) whether or not Vector excess columns
are attempted to be sorted (TRUE) or not attempted to be sorted (FALSE)}

\item{chopVectExc}{chop vector excess. FALSE(default) whether or not excess (x) elements
not found in InitOrder are removed}
}
\value{
vector sorted by InitOrder
}
\description{
\preformatted{
Performs a custom sort of a vector x.  In parameter vector InitOrder,
the elements to be sorted and the exact order is chosen.
}

\preformatted{
}

\preformatted{
Excess Vector elements are appended to the end.
(sort or not sort, CI(case insensitive) sort or non-CI sort)

For the purpose of "not sorting",
other vector elements(vector excess) that are
in the parameter vector InitOrder that are
"not found" in the input x are ignored.
}

\preformatted{
}

\preformatted{
}

\preformatted{
}

\preformatted{
}

\preformatted{
}

\preformatted{
}
}
\examples{
\dontrun{

cSort( c("a","v", "E2", "c","l", "e3" ,"h","o","date"),
   InitOrder = c("date", "o", "h", "l", "c", "v", "a"), CI = TRUE
)
[1] "date" "o"    "h"    "l"    "c"    "v"    "a"    "E2"   "e3"

cSort(c("E","B","C","D","A"), c("D","B","C"), sortVectExc = FALSE)
[1] "D" "B" "C" "E" "A"

# excess(Vector)  "G", "F"
# cSort(c("G", "D","B","C", "F"), c("E","B","C","D","A"))
[1]"B" "C" "D" "F" "G"

cSort(c("G", "D","B","C", "F"), c("E","B","C","D","A"), sortVectExc = FALSE)
[1] "B" "C" "D" "G" "F"

# other(InitOrder) ignored "F"
cSort(c("E","B","C","D","A"), c("F", "D","B","C"), sortVectExc = FALSE)
[1] "D" "B" "C" "E" "A"

}
\dontrun{
cSort(c(7,4,2,3,6), c(5,2,3,4,1))
[1] 2 3 4 6 7
}
\dontrun{
# Date examples

cSort(zoo::as.Date(6:3), zoo::as.Date(1:4))
[1] "1970-01-04" "1970-01-05" "1970-01-06" "1970-01-07"

cSort(zoo::as.Date(6:3), zoo::as.Date(1:4), sortVectExc = FALSE)
[1] "1970-01-04" "1970-01-05" "1970-01-07" "1970-01-06"

cSort(zoo::as.Date(6:3), zoo::as.Date(1:4), sortVectExc = FALSE, chopVectExc = TRUE)
[1] "1970-01-04" "1970-01-05"
}
}
\references{
\cite{Custom Sorting in R \url{https://stackoverflow.com/questions/23995285/custom-sorting-in-r}}

\cite{Case insensitive sort of vector of string in R \url{https://stackoverflow.com/questions/29890303/case-insensitive-sort-of-vector-of-string-in-r}}
}
