% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbAddKeyEM}
\alias{dbAddKeyEM}
\title{Add Key}
\usage{
dbAddKeyEM(
  conn,
  name,
  colname,
  only = FALSE,
  type = c("primary", "foreign"),
  reference,
  colref,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A connection object.}

\item{name}{A character string, or a character vector, specifying a PostgreSQL table name.}

\item{colname}{A character string specifying the name of the column to which the key will be assign; alternatively, a character vector specifying the name of the columns for keys spanning more than one column.}

\item{only}{Logical. Default is FALSE. Whether to add to apply this key just to this parent table(TRUE). Otherwise, also apply this contraint to inherited tables(FALSE).}

\item{type}{The type of the key, either "primary" or "foreign"}

\item{reference}{A character string specifying a foreign table name to which the foreign key will be associated (ignored if type == "primary").}

\item{colref}{A character string specifying the name of the primary key in the foreign table to which the foreign key will be associated; alternatively, a character vector specifying the name of the columns for keys spanning more than one column (ignored if type == "primary").}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
TRUE if the key was successfully added.
}
\description{
Add a primary or foreign key to a table column.
}
