% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbConnectEM}
\alias{dbConnectEM}
\title{Connect to the econModel database.}
\usage{
dbConnectEM(
  driver,
  user,
  password = user,
  host,
  dbname = user,
  port,
  tty,
  options,
  forceISOdate,
  connName,
  env,
  ...
)
}
\arguments{
\item{driver}{String. Defaults to getOption("econmodel_db_driver"). String.  Default is "PostgreSQL".  Currently only an implementation exists using PostgreSQL and PostgreSQL-like databases.}

\item{user}{String. Defaults to getOption("econmodel_db_user").}

\item{password}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_password") .}

\item{host}{String. Defaults to getOption("econmodel_db_host").}

\item{dbname}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_dbname").}

\item{port}{Integer. Defaults to getOption("econmodel_db_port").}

\item{tty}{Default to getOption("econmodel_db_tty").}

\item{options}{Defaults to getOption("econmodel_db_dboptions").}

\item{forceISOdate}{Logical. Default is getOption("econmodel_db_forceISOdate").}

\item{connName}{String.  Name of the database connection object.  The default is "connEM".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{...}{Dots passed.}
}
\value{
DBI connection object named "connEM" is created, connected and assigned to the environment "env".
}
\description{
Currently is only implemented to work on PostgreSQL or PostgreSQL-like databases.
}
\details{
This function tries to return a new "DBI" connection object to the environment "env".

First, this function will try to connect to the database using the getOption("econmodel_db_*") parameters listed in the "parameters" section.

If the connection can not be made, then next the connection is tried to be made using the user/password/dbname/schema of "r_user_econmodel".
Next, the user/password/database/schema of name getOption("econmodel_storage_name") will attempt to be created.
getOption("econmodel_storage_name") defaults to the lower case name of the tempdir() tail folder.
Next, the new connection object "connEM" will be returned to the "env".
Next, disconnection occurs. Connection re-occurs using the user getOption("econmodel_storage_name").

If the connection still can not be made, then next the connection is tried to be made using the user/password/dbname/schema of "postgres".
If successful, the login/user/dbname/schema (if any do not exists) of "r_user_econmodel" will attempt to be created.
Next, the user/password/database/schema of name getOption("econmodel_storage_name") will attempt to be created.
getOption("econmodel_storage_name") defaults to the lower case name of the tempdir() tail folder.  This is set in this package econModel ".unLoad" function.
Next, the new connection object "connEM" will be returned to the "env".
Next, disconnection occurs. Connection re-occurs using the user getOption("econmodel_storage_name").

If the connection still can not be made, the program stops with a message.
}
\examples{
\dontrun{
dbConnectEM()
}
}
