% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbConnectEM}
\alias{dbConnectEM}
\title{Connect to the econModel database.}
\usage{
dbConnectEM(
  driver,
  user,
  password = user,
  host,
  dbname = user,
  port,
  tty,
  options,
  forceISOdate,
  connName,
  env,
  display = TRUE,
  exec = TRUE,
  ...
)
}
\arguments{
\item{driver}{String. Defaults to getOption("econmodel_db_driver"). String.  Default is "PostgreSQL".  Currently only an implementation exists using PostgreSQL and PostgreSQL-like databases.}

\item{user}{String. Defaults to getOption("econmodel_db_user").}

\item{password}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_password") .}

\item{host}{String. Defaults to getOption("econmodel_db_host").}

\item{dbname}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_dbname").}

\item{port}{Integer. Defaults to getOption("econmodel_db_port").}

\item{tty}{Default to getOption("econmodel_db_tty").}

\item{options}{Defaults to getOption("econmodel_db_dboptions").}

\item{forceISOdate}{Logical. Default is getOption("econmodel_db_forceISOdate").}

\item{connName}{String.  Name of the database connection object.  The default is "connEM".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
DBI connection object named "connEM" is created, connected and assigned to the environment "env".
}
\description{
Currently is only implemented to work on PostgreSQL or PostgreSQL-like databases.
}
\details{
This function tries to return a new "DBI" connection object to the environment "env".

First, this function will try to connect to the database using the user name "user" specified from getOption("econmodel_db_*") parameters listed in the "parameters" section.

If the connection can be made, then next the connection is tried to be made using the user/password/dbname of "r_user_econmodel".
If the connection is made, then next the schema "r_user_econmodel" is attempted to be made..

If the connection can not be made, then next the connection is tried to be made using the user/password/dbname of "postgres".
If the connection can be made, the the user/password/dbname/schema of "r_user_econmodel" will be made.
Next, disconnect. Connect as "r_user_econmodel".
Create the user/password/dbname of "user"
Next, disconnect. Connect as "user".
Create the schema "user".
Store the new connection object connName in "env".
}
\examples{
\dontrun{
dbConnectEM()
}
}
