% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateDbaseEM}
\alias{dbCreateDbaseEM}
\title{Database Creation}
\usage{
dbCreateDbaseEM(conn, dbname, owner = dbname, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{dbname}{String. Required.  Potential database in the cluster.}

\item{owner}{String. Database owner. Defaults to parameter "dbname".}

\item{...}{Dots Passed.}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a a database in the cluster.
}
\examples{
\dontrun{
dbCreateDbaseEM(conn, dbname = "r_user_econmodel")
dbCreateDbaseEM(conn, dbname = "rtmp")
}
}
