% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateSchemaEM}
\alias{dbCreateSchemaEM}
\title{Schema Creation}
\usage{
dbCreateSchemaEM(
  connName,
  schema,
  role_specification = schema,
  grant_all_roles = schema,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{schema}{String. Required.  Potential schema in the database.}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{role_specification.}{String. The schema role specification.  Defaults to "schema".}

\item{grant_all}{vector of Strings. Roles to be GRANT ALLed to this schema.  Defaults to "schema".}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a schema in the database.
}
\examples{
\dontrun{
 dbCreateSchemaEM(schema = "r_user")
}
}
