% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateSchemaEM}
\alias{dbCreateSchemaEM}
\title{Schema Creation}
\usage{
dbCreateSchemaEM(
  conn,
  schema,
  role_specification = schema,
  grant_all_roles = schema,
  display = TRUE,
  exec = TRUE,
  ...
)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{schema}{String. Required.  Potential schema in the database.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}

\item{role_specification.}{String. The schema role specification.  Defaults to "schema".}

\item{grant_all}{vector of Strings. Roles to be GRANT ALLed to this schema.  Defaults to "schema".}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a schema in the database.
}
\examples{
\dontrun{
 dbCreateSchemaEM(conn, schema = "r_user_econmodel")
 dbCreateSchemaEM(conn, schema = "rtmp")
}
}
