% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateSchemaEM}
\alias{dbCreateSchemaEM}
\title{Schema Creation}
\usage{
dbCreateSchemaEM(conn, schema, role_specification)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{schema}{String. Required.  Potential schema in the database.}

\item{role_specification.}{String. The schema role specification.  Required.}
}
\value{
TRUE(success) or FALSE(failure)
}
\description{
Create a schema in the database.
}
\examples{
\dontrun{
 dbCreateSchemaEM(conn, schema = "r_user_econmodel",  role_specification = "r_user_econmodel")
}
}
