% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateUserEM}
\alias{dbCreateUserEM}
\title{User Creation}
\usage{
dbCreateUserEM(
  connName,
  user,
  attributes = c("LOGIN"),
  password = user,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{user}{String. Required.  Potential user in the database.}

\item{attributes}{vector of Strings. User attributes.}

\item{password}{String. Defaults to "user".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a user exists in the database.
}
\examples{
\dontrun{
dbCreateUserEM(user = "r_user", attributes = c("LOGIN", "CREATEDB", "CREATEROLE"))
}
}
