% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbCreateUserEM}
\alias{dbCreateUserEM}
\title{User Creation}
\usage{
dbCreateUserEM(
  conn,
  user,
  attributes = c("LOGIN"),
  password = user,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{user}{String. Required.  Potential user in the database.}

\item{attributes}{vector of Strings. User attributes.}

\item{password}{String. Defaults to "user".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
TRUE(success) or Error(failure)
}
\description{
Create a user exists in the database.
}
\examples{
\dontrun{
dbCreateUserEM(conn, user = "r_user_econmodel", attributes = c("LOGIN", "CREATEDB", "CREATEROLE"))
dbCreateUserEM(conn, user = "rtmp")
}
}
