% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbdfMatchColsEM}
\alias{dbdfMatchColsEM}
\title{Create New Columns In the Data.frame or On the Server Table}
\usage{
dbdfMatchColsEM(
  conn,
  name = substitute(value),
  value,
  temporary = FALSE,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A DBIConnection PostgreSQL object, as returned by DBI::dbConnect().}

\item{name}{a character string specifying a table name.}

\item{value}{a data.frame (or coercible to data.frame).}

\item{temporary}{Logical. Default is FALSE, If TRUE, then look for the "conn" object "name" in the temporary namespace.}

\item{display}{Logical. Whether to display the query (defaults to TRUE).}

\item{exec}{Logical. Whether to execute the query (defaults to TRUE).}
}
\value{
data.frame with more columns and/or re-ordered columns. The server "name" may get more columns.
}
\description{
Match the the Server/data.frame columns with each other.
Create new columns as necessary. Verify/Make the data.frame
columns order match that of the Server table order.
Inspired by the R CRAN package Caroline.
}
\examples{
\dontrun{

conn <- DBI::dbConnect(RPostgreSQL::PostgreSQL(), user = "postgres")

mtcars2 <- mtcars
mtcars2[["model"]] <- rownames(mtcars2)
mtcars2 <- DataCombine::MoveFront(mtcars2, Var = "model")
mtcars2[["vs"]] <- as.logical(mtcars2[["vs"]])
mtcars2[["gear"]] <- as.integer(mtcars2[["gear"]])
mtcars2[["carb"]] <- zoo::as.Date(mtcars2[["carb"]])
rownames(mtcars2) <- NULL
#
# Creates the table (with zero rows).
# Appends data (with the Df having the same columns that the server).
mtcars2s <- mtcars2[1:5,]
dbdfMatchColsEM(conn, name = "mtcars",  Df = mtcars2s, PartKeyDef = "LIST (gear)", PrimaryKey = c("gear", "model"), Indexes = list(gear_model_vs = c("gear", "model", "vs")))

# Appends data (with the Df having less columns that the server database).
# Those server columns, that are not found in the Df, are added to the Df.
mtcars2lDf <- mtcars2[6:10, "model", drop = F]
dbdfMatchColsEM(conn, name = "mtcars", Df = mtcars2lDf)

# Appends data (with the server database having less columns that the Df).
# Those Df columns, that are not found in the server, are added to the sever.
mtcars2lSv <- {DfNew <- mtcars2[11:15, c("model","vs", "am", "gear", "carb")]
               colnames(DfNew) <- paste0(colnames(DfNew),"_new")
               DfNew[["model"]] <- DfNew[["model_new"]]
               DfNew <- DataCombine::MoveFront(DfNew, Var = "model")
               DfNew
              }; rm(DfNew)

dbdfMatchColsEM(conn, name = "mtcars", Df = mtcars2lSv)
}
}
