% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbDisconnectEM}
\alias{dbDisconnectEM}
\title{Disconnect from and econModel database}
\usage{
dbDisconnectEM(connName, env)
}
\arguments{
\item{connName}{String.  Name of the database connection object. Optional. This may be "passed" instead of  "conn".}

\item{env}{Environment. Default is the global environment .GlobalEnv.  Location of the connection object "connName".}
}
\value{
Disconnects "conn" from the database or disconnects "connName" from the database and removes it from the environment "env".
}
\description{
Disconnects.
}
\examples{
\dontrun{
dbDisconnectEM() # default is connection variable "connEM" in the .GlobalEnv
# Tried to disconnect the R object "connEM" and remove it
# from the environment <environment: R_GlobalEnv>.

dbDisconnectEM(connName = "connEM", env = .Globalenv)
}
}
