% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbDisconnectEM}
\alias{dbDisconnectEM}
\title{Disconnect from and econModel database}
\usage{
dbDisconnectEM(connName, env, ...)
}
\arguments{
\item{connName}{String.  Required.  Name of the database connection object}

\item{env}{Environment.  Default is the global environment .GlobalEnv.  Location of the connection object "connName".}

\item{...}{Dots passed.}
}
\value{
Disconnects "connName" and removes it from the environment "env".
}
\description{
Also, removes the DBI connection object from the environment "env".
}
