% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExecuteEM}
\alias{dbExecuteEM}
\title{Executes in the  database}
\usage{
dbExecuteEM(conn, Statement, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Optional. This may be "passed" instead of  "connName".}

\item{Statement}{String. Required.  DML/DDL/DCL to execute}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
Execution of "conn" from the database or disconnects "connName" from the database and removes it from the environment "env".
}
\description{
Disconnects the R object passed "conn".
Alternately, can disconnect and remove the DBI connection object connName (default is "connEM") from the environment "env".
The function will just do one or the other, but not both.  Does NOT remove the passed R object "conn".
}
\examples{
\dontrun{
dbExecuteEM(get("connEM"), Statement = "CREATE TABLE xyz();")
}
}
