% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExistsDbaseEM}
\alias{dbExistsDbaseEM}
\title{Database Existence?}
\usage{
dbExistsDbaseEM(conn, dbname, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{dbname}{String. Required.  Potential database in the cluster.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
TRUE(exists) or FALSE(not exists)
}
\description{
Determine if a database exists in the cluster.
}
\examples{
\dontrun{
 dbExistsDbaseEM(conn, dbname = "r_user_econmodel")
 dbExistsDbaseEM(conn, dbname = "rtmp")
}
}
