% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExistsSchemaEM}
\alias{dbExistsSchemaEM}
\title{Schema Existence?}
\usage{
dbExistsSchemaEM(conn, schema, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{schema}{String. Required.  Potential schema in the database.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
TRUE(exists) or FALSE(not exists)
}
\description{
Determine if a schema exists in the database.
}
\examples{
\dontrun{
 dbExistsSchemaEM(conn, schema = "r_user_econmodel")
 dbExistsSchemaEM(conn, schema = "rtmp")
}
}
