% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbExistsUserEM}
\alias{dbExistsUserEM}
\title{User Existence?}
\usage{
dbExistsUserEM(conn, user, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection. Required.}

\item{user}{String. Required.  Potential user in the database.}

\item{...}{Dots passed.}
}
\value{
TRUE(exists) or FALSE(not exists)
}
\description{
Determine if a user exists in the database.
}
\examples{
\dontrun{
 dbExistsUserEM(conn, user = "r_user_econmodel")
 dbExistsUserEM(conn, user = "rtmp")
}
}
