% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetCurrentUserEM}
\alias{dbGetCurrentUserEM}
\title{PostgreSQL Current User}
\usage{
dbGetCurrentUserEM(connName, env, exec = TRUE, display = TRUE)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}
}
\value{
Current user
}
\description{
Get PostgreSQL CURRENT_USER.
}
\examples{
\dontrun{
dbGetCurrentUserEM()
}
}
