% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetInfoExtraEM}
\alias{dbGetInfoExtraEM}
\title{Extra Information About The RPostgreSQL Connection}
\usage{
dbGetInfoExtraEM(conn, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection}

\item{...}{Dots passed.}
}
\value{
R list of Strings of properties.
}
\description{
Adds extra information: current_schema, search_path, temp_dbname, and timeZone.
It will not report the user temporary database (temp_dbname) (if it does not yet exist).
}
\examples{
\dontrun{
dbInfoExtraEM(conn)
}
}
