% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetQueryEM}
\alias{dbGetQueryEM}
\title{Results column names are always in UPPERCASE.}
\usage{
dbGetQueryEM(conn, Statement, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by dbConnect().}

\item{...}{Other parameters passed on to methods.}

\item{statement}{a character string containing SQL.}
}
\description{
Thin wrapper over R CRAN package DBI function dbGetQuery
}
