% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbGetQueryEM}
\alias{dbGetQueryEM}
\title{Results column names are always in UPPERCASE.}
\usage{
dbGetQueryEM(
  connName,
  Statement,
  time_zone = "UTC",
  client_encoding = "UTF8",
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{Statement}{a character string containing SQL. The statement "SELECT 1;" will test if a database connection is non-expired or is-valid. Then, the returned value is either data.frame(DBISCONNECTEDEM = TRUE) or data.frame(DBISCONNECTEDEM = TRUE).}

\item{time_zone}{Execution time zone. Default is "UTC".}

\item{client_encoding}{Execution encoding. Default is "UTF8".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
SQL results.  Otherwise, an error is returned.  The results column names are always in uppercase.
}
\description{
Thin wrapper over R CRAN package DBI function dbGetQuery
}
