% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbIsConnectedEM}
\alias{dbIsConnectedEM}
\title{Test The RPostgreSQL Connection}
\usage{
dbIsConnectedEM(conn, display = TRUE, exec = TRUE, ...)
}
\arguments{
\item{conn}{PostgreSQL DBI connection.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{...}{Dots passed.}
}
\value{
TRUE(connected) or FALSE(otherwise)
}
\description{
Is the connection not expired or not valid?
}
\examples{
\dontrun{
isConnectedEM(conn)
}
}
