% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbLoginEM}
\alias{dbLoginEM}
\title{Connect to the econModel database.}
\usage{
dbLoginEM(
  driver,
  connName,
  user,
  password = user,
  host,
  dbname = user,
  port,
  tty,
  options,
  forceISOdate,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{driver}{String. Defaults to getOption("econmodel_db_driver"). String.  Default is "PostgreSQL".  Currently only an implementation exists using PostgreSQL and PostgreSQL-like databases.}

\item{connName}{String.  Name of the database connection object.  The default is "connEM".}

\item{user}{String. Defaults to getOption("econmodel_db_user").}

\item{password}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_password").}

\item{host}{String. Defaults to getOption("econmodel_db_host").}

\item{dbname}{String. Defaults to "user". If missing, then defaults to getOption("econmodel_db_dbname").}

\item{port}{Integer. Defaults to getOption("econmodel_db_port").}

\item{tty}{Default to getOption("econmodel_db_tty").}

\item{options}{Defaults to getOption("econmodel_db_dboptions").}

\item{forceISOdate}{Logical. Default is getOption("econmodel_db_forceISOdate").}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
Invisibly a DBI connection in an object named "connEM" is created, connected, and assigned to the environment "env".
}
\description{
Currently is only implemented to work on PostgreSQL or PostgreSQL-like databases.
Try to connect or try to connect as "user".
}
\details{
Upon namespace loading, the R option "econmodel_db_storage_name"
is read. From this read value the other R options are
set: "econmodel_db_user", "econmodel_db_password",
and "econmodel_db_dbname".

See these mentioned defaults and other defaults in zzz.R.

If the R option "econmodel_db_storage_name" is not set, and
thus the R options "econmodel_db_username", "econmodel_db_password",
and "econmodel_db_dbname" are not set, and the
user does not supply a "username", then "username", "password",
and "dbname" default to the lowercase name of the tempdir().
}
\examples{
\dontrun{

# Reasons for an error:
#
# No permission
# Object already exists

# a typical workflow may follow #

# To change the default user to "postgres"
getOption("econmodel_db_user")
options(econmodel_db_storage_name = "postgres")
# In rstudio dev, do
#  (1) Session -> "Restart R",
#  (2) devtools::load_all(".")
getOption("econmodel_db_user")

dbLoginEM()
dbIsConnectedEM()
dbGetCurrentUserEM()
# "postgres"
dbGetInfoExtraEM()
dbCreateUserEM(user = "r_user", attributes = c("LOGIN", "CREATEDB", "CREATEROLE"))
dbExistsUserEM(user = "r_user")
dbCreateDbaseEM(dbname = "r_user")
dbLogoutEM()

dbLoginEM(user = "r_user")
dbIsConnectedEM()
dbGetCurrentUserEM()
# "r_user"
dbGetInfoExtraEM()
dbCreateSchemaEM(schema = "r_user")
dbGetInfoExtraEM()
dbLogoutEM()

# To change the default user to "r_user"
# if in package dev, in rstudio
# (1) Session -> "Restart R",
# (2) Session -> "Restart R",
# (3) options(econmodel_db_storage_name = "r_user")
# (4) devtools::load_all(".")
# (5) getOption("econmodel_db_user")

dbLoginEM()
dbGetCurrentUserEM()
dbGetInfoExtraEM()
dbLogoutEM()

}
}
