% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbPartBoundEM}
\alias{dbPartBoundEM}
\title{From a PostgreSQL Object Get Its Partition Bounds}
\usage{
dbPartBoundEM(connName, name, env, display = TRUE, exec = TRUE)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{name}{String. Name of the PostgreSQL object.}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
vector of partiton bound list values.
}
\description{
Of a PostgreSQL inheriting partition, detect its partition bounds.
If not a "inheriting partition", an empty character vector is returned.
}
\details{
This is only designed to work on List paritions the non-expression partition bounds.
}
