% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbPartKeyColEM}
\alias{dbPartKeyColEM}
\title{From a PostgreSQL Object Get Its Partition Key Definition}
\usage{
dbPartKeyColEM(conn, DfName)
}
\arguments{
\item{conn}{PostgreSQLConnection.}

\item{DfName}{String. Name of the PostgreSQL object.}
}
\value{
vector of size 1 of the partition key definition.
}
\description{
Of a PostgreSQL  partition table, detect its partition key definition.
If not a "partition table", an empty character vector is returned.
}
\details{
This is only designed to work on List partitioned tables with the non-expression key.
Will work: "PARTITION BY LIST (c1)" is stored as "LIST (c1)".
Will not work: "PARTITION BY LIST (CAST(c1) AS INTEGER)" is stored as "LIST (((c1)::integer))".
}
