% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbRClmnsClassesEM}
\alias{dbRClmnsClassesEM}
\title{Given a Server Table, Of Its Fields, Get the Equivalent R Object classes
If a server table, of its fields,  would be demoted to be in R, then these would be the equivalent R Object classes.}
\usage{
dbRClmnsClassesEM(conn, name, temporary = FALSE, display = TRUE, exec = TRUE)
}
\arguments{
\item{conn}{A connection object.}

\item{name}{Table name string, length 1-2. Requires at least one row.}

\item{temporary}{Logical.  This is a temporary table or not.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
data.frame. Variables are "name" and "type"
}
\description{
R Classes depend upon what the server supports.
SQLite R equivalents typically are "character", "integer", and "real".
PostgreSQL R equivalents typically are "character", integer", real", "Date", (and may be others: POSIXct?).
}
