% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbServerFieldsCClassesEM}
\alias{dbServerFieldsCClassesEM}
\title{Server Side Table Column Stored Type C Classes}
\usage{
dbServerFieldsCClassesEM(
  connName,
  name,
  temporary = FALSE,
  others = FALSE,
  env,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{name}{Table name string, length 1-2.}

\item{temporary}{Logical.  This is a temporary table or not.}

\item{others}{Logical. Default is FALSE.  Using PostgreSQL, Whether or not to also return the "others" column and values  (e.g."sclass"). Values are in lowercase. Note, the returned column order may be mixed.}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
data.frame. Variables are "name" and "type" (optionally some "others"). "type" column values are lowercase.
}
\description{
Get the internal table of database server columns.
}
\details{
\preformatted{

Typical (SQLite) columns:
NAME(chr)  TYPE(chr)

Typical SQLite column TYPE values:
TYPE: character double integer

Typical PostgreSQL  columns (others = TRUE):
NAME(chr), SCLASS(chr), TYPE(chr),
LEN(int), PRECISION(int), SCALE(int) NULLOK(logi)

Typical PostgreSQL column values:
SCLASS: character double integer logical Date
TYPE: TEXT FLOAT8 INTEGER BOOL DATE
}
}
