% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbServerFieldsCClassesEM}
\alias{dbServerFieldsCClassesEM}
\title{Server Side Table Column Stored Type C Classes}
\usage{
dbServerFieldsCClassesEM(
  conn,
  name,
  temporary = FALSE,
  display = TRUE,
  exec = TRUE
)
}
\arguments{
\item{conn}{A connection object.}

\item{name}{Table name string, length 1-2.}

\item{temporary}{Logical.  This is a temporary table or not.}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}
}
\value{
data.frame. Variables are "name" and "type" (optionally some others)
}
\description{
Get the internal table of database server columns.
}
\details{
\preformatted{

Typical (SQLite) columns:
NAME(chr)  TYPE(chr)

Typical SQLite column TYPE values:
TYPE: character double integer

Typical PostgreSQL  columns:
NAME(chr), SCLASS(chr), TYPE(chr),
LEN(int), PRECISION(int), SCALE(int) NULLOK(logi)

Typical PostgreSQL column values:
SCLASS: character double integer logical Date
TYPE: TEXT FLOAT8 INTEGER BOOL DATE
}
}
