% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbSetPerformance}
\alias{dbSetPerformance}
\title{PostgreSQL Performance}
\usage{
dbSetPerformance(connName, env)
}
\arguments{
\item{connName}{String.  Name of the database connection object.  The default is "connEM".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}
}
\value{
PostgreSQL parameters are set
}
\description{
Set PostgreSQL memory parameters.
}
\examples{
\dontrun{
dbSetPerformance()
}
}
