% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbSetPerformanceEM}
\alias{dbSetPerformanceEM}
\title{PostgreSQL Performance}
\usage{
dbSetPerformanceEM(connName, env, exec = TRUE, display = TRUE)
}
\arguments{
\item{connName}{String.  Default is "connEM". Contains the name of the variable that contains the name of the "connection" in the environment "env".}

\item{env}{Environment.  Default is the .Global environment.  This is the environment to return the connection object "connEM".}

\item{exec}{Logical. Whether to execute the query (defaults to \code{TRUE}).}

\item{display}{Logical. Whether to display the query (defaults to \code{TRUE}).}
}
\value{
PostgreSQL parameters are set
}
\description{
Set PostgreSQL memory parameters.
}
\examples{
\dontrun{
dbSetPerformanceEM()
}
}
