% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dbSetPerformanceEM}
\alias{dbSetPerformanceEM}
\title{PostgreSQL Performance}
\usage{
dbSetPerformanceEM(conn)
}
\arguments{
\item{conn}{PostgreSQL DBI connection.}
}
\value{
PostgreSQL parameters are set
}
\description{
Set PostgreSQL memory parameters.
}
\examples{
\dontrun{
dbSetPerformance(get("connEM"))
}
}
