% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Df2Xts.R
\name{castDf2Xts}
\alias{castDf2Xts}
\alias{meltDf2Xts}
\alias{indName}
\alias{indName<-}
\title{Convert a Dataframe to an xts object}
\usage{
castDf2Xts(x, ValueName = NULL, IndexName = NULL, RmNA = F)

meltDf2Xts(
  x,
  IndexVar = NULL,
  ValueVar = NULL,
  AttrVar = NULL,
  FunAggr = NULL,
  DelUniqCAttr = NULL
)

indName(x)

indName(x) <- value
}
\description{
Of castDf2Xts, it is a wrapper over the R CRAN package zoocat function cast2zoocat

Of meltDf2Xts, it is a wrapper over the R CRAN package zoocat function melt

Of indName, it is a wrapper over the R CRAN package zoocat function indname

Of `indName<-`, is a wrapper over the R CRAN package zoocat function `indname<-`
}
