% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DFA}
\alias{DFA}
\title{Differences by using Absolute Change}
\usage{
DFA(x, l = 1, d = 1, ...)
}
\arguments{
\item{l}{lag}

\item{d}{differences}
}
\description{
\preformatted{
}
}
\examples{
\dontrun{

# DFA(Differences by using Absolute Change) examples

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

#' DFA(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       -3        8
1970-01-03       -2       16

#' DFA(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)), d = 2)
           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       NA       NA
1970-01-03        1        8

DFA(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)), l = 2)
           V1diff.1 V2diff.1
1970-01-01       NA       NA
1970-01-02       NA       NA
1970-01-03       -5       24
}
}
