% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{dfServerFieldsClassesEM}
\alias{dfServerFieldsClassesEM}
\title{Given an R data.frame, Get The Equivalent Database Server Column Types
If a data.frame would be promoted to be on a database server, then these would be the equivalent Server datatypes.}
\usage{
dfServerFieldsClassesEM(conn, Df)
}
\arguments{
\item{conn}{A connection object.}

\item{Df}{data.frame}
}
\value{
data.frame. Variables are "name" and "type"
}
\description{
Given an R data.frame, Get The Equivalent Database Server Column Types
If a data.frame would be promoted to be on a database server, then these would be the equivalent Server datatypes.
}
\details{
\preformatted{

typical SQLite results of TYPE values:
"TEXT"  "REAL" "INTEGER"

typical PostgreSQL results of TYPE values:
"text"  "float8" "bool"  "integer" "date"
}
}
