% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econModel.R
\docType{package}
\name{econModel}
\alias{econModel}
\alias{econModel-package}
\title{econModel: Meaningful Social and Economic Data from ALFRED and elsewhere}
\description{
Collect and display meaningful social and economic data.
}
\section{XTS Manipulation and Derivation}{

\preformatted{
explode

ifelse.xts
is.na.xts

timeEvent
lagXts
absChg
relChg
diffXts
runRanks
}
}

\section{St. Louis ALFRED (and FRED)}{

\preformatted{
fredAttributes
estimLastUpdated
vinDates
getFRED
getALFRED
getSymbols.ALFRED

toPeriod
}
}

\section{Data Combinations}{

\preformatted{
DetectFullRows
DetectOnlyNonEmptyVarsInRows
DetectEmptyRows
RemoveEmptyRows
}
}

\section{American Association of Individual Investors (AAII)  StockInvestor Pro}{

\preformatted{
getAAIISIProDate
copyAAIISIProDBFs
}
}

\section{Willem Ligtenberg and Andre Mikulec}{

\preformatted{
unlockEnvironment
forceAssignInNamespace
lsNamespaceInfo
AllInfoNS
}
}

\section{Miscellaneous}{

\preformatted{
cSort
}
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/AndreMikulec/econModel/issues}
}

}
\author{
\strong{Maintainer}: Andre Mikulec \email{Andre_Mikulec@Hotmail.com}

}
