% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{estimLastUpdated}
\alias{estimLastUpdated}
\title{Estimate the Time Distance Between Observations Dates/Times and Their Corresponding LastUpdated(Publishing) Dates/times}
\usage{
estimLastUpdated(
  x,
  Calendar = "UnitedStates/GovernmentBond",
  Frequency,
  LastUpdated,
  LastOfDateRange = NULL
)
}
\arguments{
\item{x}{Date times of observation dates}

\item{Calendar}{Default is "UnitedStates/GovernmentBond". Calendar to use.  See ?? RQuantLib::Calendars}

\item{Frequency}{No default (required). Values can be Daily(not implemented yet), Weekly(not implemented yet), Monthly, and Quarterly}

\item{LastUpdated}{No default (required).  Date time of the published date of the newest(latest) observation}

\item{LastOfDateRange}{Default NULL.  If not present, this value will be taken from the last observation of x.  This parameter value represents the earliest of the time between an observation and its corresponding LastUpdated Date time.}
}
\value{
vector of Date times
}
\description{
Time between LastOfDateRange (or the last Date time observation of x) and LastUpdated is used to estimate
}
\examples{
\dontrun{

Title:               Unemployment Rate
Series ID:           UNRATE
Source:              U.S. Bureau of Labor Statistics
Release:             Employment Situation
Seasonal Adjustment: Seasonally Adjusted
Frequency:           Monthly
Units:               Percent
Date Range:          1948-01-01 to 2020-10-01
Last Updated:        2020-11-06 7:47 AM CST

# LastUpdated (updated) during the first Friday of every month
# if that "first Friday" is not available, then the
# the updated occurs at the next Friday
#
# Schedule of Releases for the Employment Situation
https://www.bls.gov/schedule/news_release/empsit.htm

# PAY & LEAVE
https://www.opm.gov/policy-data-oversight/pay-leave/federal-holidays/#url=2020

atr <- fredAttributes("UNRATE")

library(quantmod)

getSymbols("UNRATE", src = "FRED")

estimLastUpdated(index(getSymbols("UNRATE")),
  Frequency = atr$Frequency,
  LastUpdated = atr$LastUpdated,
  LastOfDateRange = tail(strsplit(atr$DateRange, " to "),1)
  )
}
}
