% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{explodeXts}
\alias{explodeXts}
\title{expland out a two dimensional object}
\usage{
explodeXts(
  xTs1 = NULL,
  xTs2 = NULL,
  Fun = NULL,
  Whiches = NULL,
  AltName = NULL,
  Prefix = NULL,
  FixedSep = NULL,
  quote = FALSE,
  envir = parent.frame(2),
  ...
)
}
\arguments{
\item{xTs1}{xts object}

\item{xTs2}{xts object}

\item{Fun}{function name in the "bare" or in literal quotes("")}

\item{Whiches}{list of possible varying parameters that are expanded
to all possible combinations by expand.grid}

\item{AltName}{string alternate name for "Fun"}

\item{Prefix}{boolan default is FALSE.  TRUE would place the column meta before the column name.}

\item{FixedSep}{string divider of meta items}

\item{quote}{boolean passed to package DescTools function DoCall}

\item{envir}{calling environment}
}
\value{
new xts object of new derived columns
}
\description{
\preformatted{

From one(1) or two(2) two dimension objects, apply a function(3) upon the two objects.

Return a two dimensional object of with a column name derived from the three.

Functions are meant to be many of the functions from R CRAN packages package TTR and PerformanceAnalytics

NOTE: if any xTs2, then xTs1 and xTs2 are paired/matched column position to column position.

This the "multivariate form" of the "single variate form" object TTR function data generator from the web page "Time series cross-validation 5"

}
}
\examples{
\dontrun{

library(quantmod)
ibm <- getSymbols("IBM", from = "1970-01-01", to = "1970-01-13", auto.assign = FALSE)

explodeXts(ibm[,c("IBM.Open")], Fun = "TTR::SMA", Whiches = list(n = 2:3))
# IBM.Open.TTR.SMA.n.2 IBM.Open.TTR.SMA.n.3
# 1970-01-02                   NA                   NA
# 1970-01-05               18.262                   NA
# 1970-01-06               18.356               18.312
# 1970-01-07               18.419               18.379
# 1970-01-08               18.431               18.425
# 1970-01-09               18.456               18.446
# 1970-01-12               18.463               18.454

  explodeXts(ibm[,c("IBM.Open","IBM.Close")], Fun = "TTR::SMA", Whiches = list(n = 2:3))
#
#            IBM.Open.TTR.SMA.n.2 IBM.Close.TTR.SMA.n.2 IBM.Open.TTR.SMA.n.3 IBM.Close.TTR.SMA.n.3
# 1970-01-02                   NA                    NA                   NA                    NA
# 1970-01-05               18.262                18.325                   NA                    NA
# 1970-01-06               18.356                18.419               18.312                18.358
# 1970-01-07               18.419                18.431               18.379                18.425
# 1970-01-08               18.431                18.456               18.425                18.446
# 1970-01-09               18.456                18.463               18.446                18.454
# 1970-01-12               18.463                18.419               18.454                18.438
}
}
\references{
\cite{Zachary Mayer, "Time series cross-validation 5"
\url{http://www.r-bloggers.com/time-series-cross-validation-5/}
\url{http://moderntoolmaking.blogspot.com/2013/01/time-series-cross-validation-5.html}
}
}
