% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{fetchInternet}
\alias{fetchInternet}
\title{Scrape the Internet}
\usage{
fetchInternet(
  x,
  Structure = "Page",
  Collapse = "\\n",
  encode = T,
  Message = paste0("function ", "fetchInternet"),
  conClose = T,
  ...
)
}
\arguments{
\item{x}{URL}

\item{Collapse}{String. Default is "\n". If Structure == "Page" then separate the lines by "\n". Alternately, separate the lines by a single character Collapse.  Another choice may be "". If Structure != "Page", then this parameter is ignored.}

\item{encode}{Logical. Default is TRUE. Perform URLencoding upon URL.}

\item{Message}{String. Default is paste0("function ", "fetchInternet").  Message is appended to the user agent.}

\item{conClose}{Logical.  Default is TRUE. After done with work, close the curl connection.  Otherwise, do not close: No closing is needed inside the foreach::foreach construct.}

\item{...}{Dots. Passed to URLencode}

\item{Stucture}{String. Default is "Page". Return the data as a page.  Other is "Lines" to return the result as a group of lines.}
}
\value{
The one element page or many vectors of lines
}
\description{
Returned data is either a page or a collection of vector lines
}
\examples{
\dontrun{

writeLines(fetchInternet("https://fred.stlouisfed.org/data/RECPROUSM156N.txt"))
Title:               Smoothed U.S. Recession Probabilities
Series ID:           RECPROUSM156N
Source:              Piger, Jeremy Max, Chauvet, Marcelle
Release:             U.S. Recession Probabilities
Seasonal Adjustment: Not Seasonally Adjusted
Frequency:           Monthly
Units:               Percent
Date Range:          1967-06-01 to 2020-09-01
Last Updated:        2020-11-02 7:01 AM CST

fetchInternet("https://fred.stlouisfed.org/data/RECPROUSM156N.txt", Structure = "Lines")

}
}
