% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{fredAttributes}
\alias{fredAttributes}
\title{Download attributes of Times Series of Federal Reserve Economic Data - FRED(R)}
\usage{
fredAttributes(Symbol, DataSheet = F)
}
\arguments{
\item{Symbol}{specifying the name of the symbol to be queried}

\item{DataSheet}{Logical. Default is FALSE.  Do not return the data payload. Otherwise, if TRUE, then also return the data payload of all of the data in an R attribute named "DataSheet".}
}
\value{
data.frame of attributes
}
\description{
Data typically includes the Title, Series ID, Source, Release, Seasonal Adjustment, Frequency, Units, Date Range, Last Updated, and Notes.
}
\examples{
\dontrun{

# Smoothed U.S. Recession Probabilities
# Source: Piger, Jeremy Max, Chauvet, Marcelle
# https://fred.stlouisfed.org/data/RECPROUSM156N.txt
#
print(str(fredAttributes("RECPROUSM156N")))

'data.frame':	1 obs. of  10 variables:
$ Title             : chr "Smoothed U.S. Recession Probabilities"
$ SeriesID          : chr "RECPROUSM156N"
$ Source            : chr "Piger, Jeremy Max, Chauvet, Marcelle"
$ Release           : chr "U.S. Recession Probabilities"
$ SeasonalAdjustment: chr "Not Seasonally Adjusted"
$ Frequency         : chr "Monthly"
$ Units             : chr "Percent"
$ DateRange         : chr "1967-06-01 to 2020-09-01"
$ LastUpdated       : chr "2020-11-02 7:01 AM CST"
$ Notes             : chr "Smoothed recession probabilities for the United"

tail(attr(fredAttributes("RECPROUSM156N", DataSheet = T), "DataSheet"))
RECPROUSM156N
2020-04-01        100.00
2020-05-01         37.85
2020-06-01         37.85
2020-07-01         47.92
2020-08-01        100.00
2020-09-01        100.00
}
}
