% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.ALFRED.R
\name{getSymbols.ALFRED}
\alias{getSymbols.ALFRED}
\title{Download Federal Reserve Economic Data - ALFRED(R)}
\usage{
getSymbols.ALFRED(
  Symbols,
  env,
  return.class = "xts",
  look.back = 3,
  vintages.per.query = 12,
  fullOldestVintageData = 12,
  datasheet = F,
  allowParallel = F,
  ...
)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{class of returned object}

\item{look.back}{from the 'Last Updated' date try to peek back in time to the 1st vintage with a published tail 'Date Range' date that is within variable 'look.back' periods. If the periodicy is "day" and, just after a three(3) day holiday weekend, to reach back from a Tuesday to a Friday, parameter look.back is increased to a minimum value of 4.  Default is 3.  Increase this value if much time exists between the tail date of 'Date Range' and the 'Last Updated' date: meaning zero(0) observations exist in the look.back period.  The R CRAN package xts function periodicity determines the period of time.  This function is meant to minimize CPU and disk I/O.}

\item{vintages.per.query}{number of vintages per HTTPS GET. A.k.a the number of vintages per sheet.   Default is 12.  Common maximum is 12. Value can be "Max". Practical experience has performed with 192.   The maximum may be different during different times of the day or night. This parameter exists to enhance performance by limiting the number of trips to the server.}

\item{fullOldestVintageData}{if TRUE, then also return the oldest vintage data and keep(prepend) its data.  Default is FALSE. Useful when 'as much data as possible' is important.}

\item{datasheet}{if TRUE, then also return all of the vintages in an xts attribute 'datasheet'. Default is FALSE.  Useful for debugging.}

\item{allowParallel}{if TRUE, then collect groups of 'sheets of vintages.per.query vintages' in parallel.  Default is FALSE.  (Improved) performance will vary: this is more useful on (more data points) weekly data or daily data. Because this is a server side activity, the number of parallel processes does NOT depend on the local machine CPUs.}

\item{...}{additional parameters}
}
\description{
R access to over 11,000 data series accessible via the St. Louis Federal Reserve Bank's ALFRED system.
}
\details{
Downloads Symbols to specified env from 'research.stlouisfed.org'. This method is not to be called directly, instead a call to R CRAN packages quantmod function getSymbols(Symbols,src='ALFRED') will in turn call this method.

The St. Louis Federal Reserve Bank's FRED allows a specific data series to be revised(overwritten) with different data while using the same published tail date of 'Date Range'.  This revision is called the 2nd vintage. This function collects and displays data from the 1st vintage that only contains non-revised data (that means the original data) per published tail date of 'Date Range'.
}
\examples{
\dontrun{

library(quantmod)

# Smoothed U.S. Recession Probabilities (RECPROUSM156N)
# Source: Piger, Jeremy Max, Chauvet, Marcelle
# https://fred.stlouisfed.org/data/RECPROUSM156N.txt
#
getSymbols("RECPROUSM156N", src =   "FRED")
[1] "RECPROUSM156N"

getSymbols("RECPROUSM156N", src = "ALFRED", look.back = 4)

Read 2002 items
Beginning Vintage: . . . 2012-09-04
Beginning Vintage: . . . 2013-09-03
Beginning Vintage: . . . 2014-09-01
Beginning Vintage: . . . 2015-09-01
Beginning Vintage: . . . 2016-09-01
Beginning Vintage: . . . 2017-09-01
Beginning Vintage: . . . 2018-09-03
Beginning Vintage: . . . 2019-10-01
Beginning Vintage: . . . 2020-10-01
[1] "RECPROUSM156N.vin"

# Note, the distance between the graphic date and the 'Last Updated' date
# is about two(2) months, so one may shift the graphic to the right
# by two(2) months to get the real story.  (The now proposed shift
# is not done/shown here.)
dygraphs::dygraph(merge(RECPROUSM156N, RECPROUSM156N.vin, join = "inner"))

# if too much time in periods exists between the
# tail date of  the 'Date Range' and 'Last Updated' date,
# then increase look.back from three(3) (default) to four(4) or greater.
getSymbols("RECPROUSM156N", src = "ALFRED", look.back = 4)

# for debugging, include the vintages
getSymbols("RECPROUSM156N", src = "ALFRED", look.back = 4, datasheet = T)
View(data.frame(xtsAttributes(RECPROUSM156N.vin)$datasheet))

# prepend (include) the data of the very first vintage in the head of the data
getSymbols("RECPROUSM156N", src = "ALFRED", look.back = 4, fullOldestVintageData = T)

# Use R CRAN package doParallel to query simultaneously
getSymbols("RECPROUSM156N", src = "ALFRED", look.back = 4, allowParallel = T)

# quarterly
# very large (that was back-loaded from 1991)
# Gross Domestic Product (GDP)
getSymbols("GDP", src = "ALFRED")
# noticably faster
getSymbols("GDP", src = "ALFRED", allowParallel = T)

# weekly
# Weekly, Ending Friday (reported during the following Wednesday)
# Chicago Fed National Financial Conditions Index (NFCI)
getSymbols("NFCI", src = "ALFRED")

# daily
# Effective Federal Funds Rate (EFFR)
getSymbols("EFFR", src = "ALFRED")
# better
getSymbols("EFFR", src = "ALFRED", vintages.per.query = 192)

}
}
