% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{getYieldCurve}
\alias{getYieldCurve}
\title{Download US Treasury Yield Curve Data}
\usage{
getYieldCurve(
  base = "http://data.treasury.gov/feed.svc/DailyTreasuryYieldCurveRateData",
  year = NULL,
  month = NULL
)
}
\arguments{
\item{base}{String. Required. The base URL for the data service, defaulting to http://data.treasury.gov/feed.svc/DailyTreasuryYieldCurveRateData. If the month or year arguments are not NULL, then the function modifies this URL to parameterize the download request.}

\item{month}{The desired month number or NULL for all months (default).}

\item{yearThe}{desired year number or NULL for all years (default)}
}
\value{
xts object
}
\description{
Downloads US Treasury yield curve data from the US Treasury web site.
}
\examples{
\dontrun{
xlist <- getYieldCurve(base = "http://data.treasury.gov/feed.svc/DailyTreasuryLongTermRateData?$filter=month(QUOTE_DATE)\%20eq\%2012\%20and\%20year(QUOTE_DATE)\%20eq\%202020")
}
}
