% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LD}
\alias{LD}
\title{Lead an xts Object}
\usage{
LD(x, k = 1, na.pad = TRUE, ...)
}
\arguments{
\item{x}{xts object}

\item{k}{choose 1 or greater  to look into the future}

\item{na.pad}{as lag.xts}

\item{...}{dots passed to lag.xts}
}
\description{
\preformatted{

This does not complain when: any(abs(k) > NROW(xTs)):
zoo_lag can not and will not handle. So k's are eliminated
beforehand.

}
}
\examples{
\dontrun{

# LD(lagXts leading 1) example

LD(IBM.Open.TTR.SMA.n.2, k = 1)
           IBM.Open.TTR.SMA.n.2.lead.1
1970-01-02                    18.26250
1970-01-05                    18.35625
1970-01-06                    18.41875
1970-01-07                    18.43125
1970-01-08                    18.45625
1970-01-09                    18.46250
1970-01-12                          NA
}
}
