% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LG}
\alias{LG}
\title{Lag an xts Object}
\usage{
LG(x, k = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{k}{choose 0 or 1 or greater to peer into the current and the past}
}
\value{
xts object
}
\description{
\preformatted{
This does not complain when: any(abs(k) > NROW(xTs)):
zoo_lag can not and will not handle. So k's are eliminated
beforehand.
}
}
\examples{
\dontrun{

# LG(lag 1) example

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

LG(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1lg.1 V2lg.1
1970-01-01     NA     NA
1970-01-02      1      8
1970-01-03     -2     16
}
}
