% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{newColName}
\alias{newColName}
\title{generate a good column name}
\usage{
newColName(
  x = NULL,
  Fun = NULL,
  isCharFun = NULL,
  x1 = NULL,
  x2 = NULL,
  WhichCombo = NULL,
  AltName = NULL,
  Prefix = NULL,
  FixedSep = NULL
)
}
\arguments{
\item{x}{object with the old column names}

\item{Fun}{literally written function or the function name inside a string}

\item{isCharFun}{hint as to Fun being a function or the function name inside of a string}

\item{x1}{object providing part of the new colum name}

\item{x2}{object providing (yet another) part of the the new column name}

\item{WhichCombo}{list with named elements and their values to (eventually) become part of the new column name}

\item{AltName}{column new root name.  Default is NULL. Unless a string is provided then the root name will not be replaced.}

\item{Prefix}{place the new addition to the column name at the front instead of the end. Default is NULL. Internally the default is FALSE.}

\item{FixedSep}{replacement for "[.]|::" that was found in the function name}
}
\value{
object with the new column names
}
\description{
\preformatted{

Given column names of one or two two dimensional objects,
each with just one single column,  and a function and its arguments,
generate a new column name and return the original 'one colun' two dimensional object

}
}
