% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBasis.R
\name{NVAR}
\alias{NVAR}
\alias{NVAR.character}
\alias{NVAR.integer}
\alias{NVAR.numeric}
\alias{NVAR.complex}
\alias{NVAR.raw}
\alias{NVAR.xts}
\alias{NVAR.default}
\title{number of variables}
\usage{
NVAR(x = NULL)

\method{NVAR}{character}(x = NULL)

\method{NVAR}{integer}(x = NULL)

\method{NVAR}{numeric}(x = NULL)

\method{NVAR}{integer}(x = NULL)

\method{NVAR}{complex}(x = NULL)

\method{NVAR}{raw}(x = NULL)

\method{NVAR}{xts}(x = NULL)

\method{NVAR}{default}(x = NULL)
}
\arguments{
\item{x}{object}
}
\value{
integer number of variables
}
\description{
\preformatted{

NCOL wrongly returns value one(1)
on non-data.frame 2nd dimension objects
with a 2nd dimension size of zero(0).

}
}
\examples{
\dontrun{
NVAR(character())
# [1] 0
}
\dontrun{
NVAR(integer())
# [1] 0
}
\dontrun{
NVAR(numeric())
# [1] 0
}
\dontrun{
NVAR(integer())
# [1] 0
}
\dontrun{
NVAR(complex())
# [1] 0
}
\dontrun{
NVAR(raw())
# [1] 0
}
\dontrun{
# > library(xts)
# > NVAR(xts(, as.Date("1970-01-12")))
# [1] 0
}
}
