% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBasis.R
\name{NVAR}
\alias{NVAR}
\alias{NVAR.default}
\alias{NVAR.xts}
\title{Number of Variables}
\usage{
NVAR(x)

\method{NVAR}{default}(x)

\method{NVAR}{xts}(x)
}
\arguments{
\item{x}{object}
}
\value{
integer number of variables
}
\description{
\preformatted{
NCOL wrongly returns value one(1)
on non-data.frame 2nd dimension objects
with a 2nd dimension size of zero(0).

This function NVAR correctly determines the number
of variables (NVAR) in an object.

R S3 class implementations exist for classes: xts

Contributions are welcome.
}

\preformatted{
}
}
\examples{
\dontrun{

# examples

NVAR(NULL)
[1] 0

NVAR(numeric())
[1] 0

NVAR(integer())
[1] 0

NVAR(character())
[1] 0
}
\dontrun{

# xts example

library(xts)
NVAR(xts(, as.Date("1970-01-01")))
[1] 0
}
}
