% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{stressIndex}
\alias{stressIndex}
\title{Stress Index}
\usage{
stressIndex(
  ValData,
  ValDates,
  proc = "pca",
  TrainData,
  TrainDates,
  TestData,
  TestDates,
  ...
)
}
\arguments{
\item{ValDates}{List of (or vector of) pairs of vectors of begin and end date-time ranges.  Required (if not all of the ValidationData is to be used).}

\item{proc}{String. Required. Default is "pca" (Principal Component Analysis)}

\item{TrainData}{xts object. Required.}

\item{TrainDates}{List of (or vector of) pairs of vectors of begin and end date-time ranges.  Required (if not all of the ValidationData is to be used).}

\item{TestData}{xts object. Required (if any Testing (Tuning) is required.}

\item{TestDates}{List of (or vector of) pairs of vectors of begin and end date-time ranges.  Required (if not all of the ValidationData is to be used).}

\item{ValData.}{xts object. Required. This is the input used for prediction.}
}
\value{
Prediction (or Index)
}
\description{
Uses a precedure (e.g. Principal Components) to create a "stress index".
}
\examples{
\dontrun{
}
}
