% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPeriod.R
\name{toPeriod}
\alias{toPeriod}
\title{Convert a Date time series to a series of another frequency}
\usage{
toPeriod(
  x,
  Period = "months",
  PeriodEnd = NULL,
  fillMissing = T,
  fillMissingBy = "days",
  Calendar = "UnitedStates/GovernmentBond"
)
}
\arguments{
\item{x}{xts object}

\item{PeriodEnd}{String. Default is NULL meaning use the expected period end.  If Period is "weeks", this value can be  (1 - Monday, 2 - Tuesday, ... 7 - Sunday).  See ? DescTools::Weekday}

\item{Calendar}{Default is "UnitedStates/GovernmentBond". Calendar to use.  See ?? RQuantLib::Calendars.  NOT IMPLEMENTED YET.}

\item{period}{Period to convert to. Default is "months".
See ? seq.POSIXt: "secs", "mins", "hours", "days", "weeks", "months", "quarters" or "years"
This is the aggregation (summary)}

\item{fillInterior}{Logical. Default is TRUE. Created sub-Period data points.  The default is "days". NOT IMPLEMENTED YET.}

\item{fillInteriorBy}{String. If fillInterior is TRUE, then default is "days". Put in sub-Period data. NOT IMPLEMENTED YET.}
}
\value{
modified
}
\description{
fillInteriorBy data values are done by "last observation carried forward".
NOT IMPEMENTED YET
}
\examples{
\dontrun{
x <- xts(c(2,362), zoo::as.Date(c(2,362)))
toPeriod(x,  Period="months")
toPeriod(x,  Period="weeks")
}
}
