% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/To.Period.R
\name{toPeriod}
\alias{toPeriod}
\title{Convert time series data to an monthly OHLC series}
\usage{
toPeriod(
  x,
  period = "months",
  indexAt = "firstof",
  drop.time = TRUE,
  name = NULL,
  return.class = "xts",
  fillMissingDates = T,
  fillMissingData = F,
  Alignment = "regular",
  ...
)
}
\arguments{
\item{x}{? xts::to.monthly and the index(x) class must
have an S3 method for "seq(by = "months")".
Classes "Date" and POSIXt are provided  S3 methods in package "base".}

\item{period}{period to convert to. Default is "months".
See ? xts::to.period: "seconds", "minutes", "hours", "days", "weeks", "months", "quarters", and "years"
This is the aggregation (summary)}

\item{indexAt}{"firstof"(default). Different from xts::to.monthly.
See ? xts::to.period: 'yearmon', 'yearqtr', 'firstof', 'lastof', 'startof', or 'endof'.
This is the data item date time type. ERROR: startof and endof return the same dates (endof dates).}

\item{drop.time}{See ? xts::to.period}

\item{name}{See ? xts::to.period}

\item{return.class}{See ? quantmod::getSymbols.FRED: ts, zoo, xts, or timeSeries}

\item{fillMissingDates}{TRUE(default). If indexAt is one of "firstof"
or "lastof", then fill in missing month dates, if any.
Otherwise FALSE, just pass throught to xts::to.monthly.}

\item{fillMissingData}{FALSE(default).  Estimate value.  See timeSeries::na.omit .Meant for test data (not prediction data and not validation data).  If TRUE, then the parameter fillMissingDates become TRUE. NOT IMPLEMENTED YET.}

\item{...}{dots See. ? xts::to.monthly}

\item{Alignment.}{Does the x period start date match the calendar period start date? If TRUE, then Alignment is "regular." Default is "regular".  Otherwise, choose "irregular". NOT IMPLEMENTED YET.}
}
\value{
See. ? xts::to.monthly
}
\description{
\preformatted{

This is a wrapper around xts to.period.
In the case of period="months"
Months between the first date and the last
date will have and entry filled in.
It will get the "firstof" or "lastof"
months between dates as the same in the call.

x note, "firstof", "lastof" broken in indexAt
x should have generated _new_ dates (It did not.)
x the first time of the period, the last time of the period
x use (program in) zoo yearmon and yearqtr to generate these
x x <- xts(2:363, zoo::as.Date(1:362))

# weirdness or bugs of xts::to.period
# indexAt "startof" and "endof" return the same dates ("endof" dates)
# if indexAt == "firstof" then the returned index class is "POSIXct" "POSIXt"
# if indexAt == "lastof" then the returned index class is "Date"

}
}
\examples{
\dontrun{
x <- xts(2:362, zoo::as.Date(2:362))
x <- xts(c(2,362), zoo::as.Date(c(2,362)))
x <- xts(c(2,362), as.POSIXct(zoo::as.Date(c(2,362))))
toPeriod(x, indexAt = "firstof")
toPeriod(x, indexAt = "lastof")
toPeriod(x, indexAt = "yearmon")
toPeriod(x, indexAt = "yearqtr")
}
}
