% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unlockEnvironment}
\alias{unlockEnvironment}
\title{unlock an R package on the R search() path}
\usage{
unlockEnvironment(env)
}
\arguments{
\item{env}{environment}
}
\description{
Note: R package base is unlocked by design
}
\examples{
\dontrun{
library(econModel)
unlockEnvironment(as.environment(2))
}
}
\references{
\cite{unlockEnvironment function
\url{https://github.com/openanalytics/Rango/blob/adc99e077b71c8c6826cabb7ff1266050898718a/Rango/src/unlockEnvironment.cpp}
}

\cite{envir.c in the R 2.15.1 source (still in R 3.3.1)
\url{https://github.com/SurajGupta/r-source/blob/a28e609e72ed7c47f6ddfbb86c85279a0750f0b7/src/main/envir.c}
}
}
\author{
Willem Ligtenberg
}
