% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmerAssocIndividInvestorsAAII.R
\name{viewSIPRO}
\alias{viewSIPRO}
\title{Load and View AAII StockInvestor Pro FoxPro/DBase (DBF) Files}
\usage{
viewSIPRO(Collection, Source = "Install", SubDir, ...)
}
\arguments{
\item{Collection}{String.  Some of "Base", "Prices", "Sheets", or "Dictionary". Required.}

\item{Source}{String.  One of "Install" or "Repository".}

\item{SubDir}{String.  Single directory containing the "DBF" files. If "Source" = "Repository", then Required. Otherwise, ignored.}
}
\value{
String. Name of the data.frame loaded. data.frames are loaded into the environment.  Attempted to be started is View() upon each data.frame.
}
\description{
Creates a data.frame from each  file.  Loads those data.frames into the .Global environment.  Calls View() upon each data.frame (so the user may see the data).
}
\examples{
\dontrun{
viewSIPRO("Base")
viewSIPRO("Base", Source = "Repository", SubDir = "C:\\\\DATA\\\\AAIISIPRO\\\\MONTHDATE\\\\18565")
}
}
